PR fix
- No subfolder was generated for routes without Prefix, but with matching base name.
- Additional test.
