--- resources/assets/assets/index-DpRWb6Hx.js
@@ -1,9 +1,0 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/vendor-utils-DjfM5PxJ.js","assets/rolldown-runtime-CkqCuyE9.js"])))=>i.map(i=>d[i]);
-import{r as e}from"./rolldown-runtime-CkqCuyE9.js";import{_ as t,a as n,c as r,d as i,f as a,g as o,h as s,i as c,l,m as u,n as d,o as f,p,r as m,s as h,t as g,u as _,v}from"./vendor-react-emvhMHcg.js";import{C as y,t as b}from"./vendor-B5MdJxr-.js";import{a as x,c as S,i as C,n as w,o as T,r as E,s as D,t as O}from"./vendor-d3-BerEHusR.js";import{t as k}from"./vendor-utils-DjfM5PxJ.js";(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var A=e(v(),1),j=t();function M(){let[e,t]=(0,A.useState)(null),[n,r]=(0,A.useState)(!0),[i,a]=(0,A.useState)(null);return(0,A.useEffect)(()=>{fetch(`/_laravel-brain/.graph-manifest.json`).then(e=>{if(!e.ok)throw Error(`HTTP ${e.status}`);return e.json()}).then(e=>{t(e),r(!1)}).catch(e=>{a(e.message),r(!1)})},[]),{manifest:e,loading:n,error:i}}function N(e){return[...e.nodes.map(e=>({data:{id:e.id,label:e.label,type:e.type,...e.data,metrics_cc:e.data?.metrics?.cyclomaticComplexity??0}})),...e.edges.map(e=>({data:{id:e.id,source:e.source,target:e.target,label:e.label,type:e.type}}))]}function P(){let[e,t]=(0,A.useState)({data:null,loading:!1,error:null}),n=(0,A.useRef)(new Map),r=(0,A.useRef)(null),i=(0,A.useCallback)(e=>{if(r.current===e)return;r.current=e;let i=n.current.get(e);if(i){t({data:i,loading:!1,error:null});return}t(e=>({...e,loading:!0,error:null})),fetch(`/_laravel-brain/`+e).then(e=>{if(!e.ok)throw Error(`HTTP ${e.status}`);return e.json()}).then(i=>{n.current.set(e,i),r.current===e&&t({data:i,loading:!1,error:null})}).catch(n=>{r.current===e&&t({data:null,loading:!1,error:n.message})})},[]);return{state:e,elements:(0,A.useMemo)(()=>e.data?N(e.data):[],[e.data]),load:i}}function F(e,t=300){let[n,r]=(0,A.useState)(t),[i,a]=(0,A.useState)(e);return e!==i&&(a(e),r(t)),(0,A.useEffect)(()=>{if(n>=e.length)return;let t=window,i=(t.requestIdleCallback?t.requestIdleCallback.bind(t):e=>setTimeout(()=>e({didTimeout:!1,timeRemaining:()=>0}),100))(()=>{r(t=>Math.min(t+200,e.length))});return()=>{t.cancelIdleCallback?t.cancelIdleCallback(i):clearTimeout(i)}},[n,e.length]),(0,A.useMemo)(()=>e.length<=t?e:e.slice(0,n),[e,n,t])}function I(){let[e,t]=(0,A.useState)(()=>localStorage.getItem(`lb-theme`)??`dark`);return(0,A.useEffect)(()=>{document.documentElement.setAttribute(`data-theme`,e),localStorage.setItem(`lb-theme`,e)},[e]),{theme:e,toggle:()=>t(e=>e===`dark`?`light`:`dark`)}}var ee=e(k(),1),L={route:`#4CAF50`,middleware:`#FF9800`,controller:`#2196F3`,livewire_component:`#FB7185`,action:`#03A9F4`,service:`#9C27B0`,validation_request:`#0d9488`,model:`#F44336`,event:`#FFD600`,job:`#607D8B`,command:`#14b8a6`,channel:`#8b5cf6`,schedule:`#f97316`,view:`#ec4899`,mail:`#f472b6`,notification:`#db2777`,enum:`#0ea5e9`,interface:`#38bdf8`,trait:`#a78bfa`,abstract_class:`#94a3b8`,service_provider:`#ca8a04`,facade:`#00BCD4`,filament_panel:`#7C3AED`,filament_resource:`#A855F7`,filament_page:`#C084FC`,filament_page_method:`#E879F9`,filament_widget:`#06B6D4`,filament_relation_manager:`#0891B2`},R={route:`#2e7d32`,middleware:`#e65100`,controller:`#1565c0`,livewire_component:`#e11d48`,action:`#0277bd`,service:`#6a1b9a`,validation_request:`#0f766e`,model:`#c62828`,event:`#b45309`,job:`#37474f`,command:`#0d7d6e`,channel:`#5b21b6`,schedule:`#c2410c`,view:`#be185d`,mail:`#c026d3`,notification:`#9d174d`,enum:`#0369a1`,interface:`#0c4a6e`,trait:`#5b21b6`,abstract_class:`#64748b`,service_provider:`#a16207`,facade:`#00838f`,filament_panel:`#5b21b6`,filament_resource:`#7e22ce`,filament_page:`#9333ea`,filament_page_method:`#a21caf`,filament_widget:`#0369a1`,filament_relation_manager:`#075985`},z={route:`#0C1A0C`,middleware:`#1C1408`,controller:`#08141C`,livewire_component:`#1C0711`,action:`#07151D`,service:`#150C1C`,validation_request:`#042f2e`,model:`#1C0C0C`,event:`#1C1A08`,job:`#0D1113`,command:`#061514`,channel:`#110c1c`,schedule:`#1c1008`,view:`#1c0a14`,mail:`#1c0f18`,notification:`#1c0510`,enum:`#071318`,interface:`#081420`,trait:`#140822`,abstract_class:`#0f172a`,service_provider:`#422006`,facade:`#001F28`,filament_panel:`#150C2A`,filament_resource:`#1A0C26`,filament_page:`#1E0F2E`,filament_page_method:`#240E30`,filament_widget:`#071A1E`,filament_relation_manager:`#06161A`},B={route:`#f0fdf4`,middleware:`#fff7ed`,controller:`#eff6ff`,livewire_component:`#fff1f2`,action:`#e0f7fa`,service:`#fdf4ff`,validation_request:`#ccfbf1`,model:`#fff1f2`,event:`#fefce8`,job:`#f1f5f9`,command:`#f0fdfa`,channel:`#f5f3ff`,schedule:`#fff7ed`,view:`#fdf2f8`,mail:`#fce7f3`,notification:`#fce7f3`,enum:`#f0f9ff`,interface:`#ecfeff`,trait:`#f5f3ff`,abstract_class:`#f1f5f9`,service_provider:`#fef9c3`,facade:`#e0f7fa`,filament_panel:`#f5f3ff`,filament_resource:`#faf5ff`,filament_page:`#fdf4ff`,filament_page_method:`#fef0ff`,filament_widget:`#ecfeff`,filament_relation_manager:`#e0f2fe`},te=`#8B6FE8`,ne=[{label:`Low`,min:1,max:5,fill:`#0d2e1a`,border:`#4ade80`},{label:`Moderate`,min:6,max:10,fill:`#2e2200`,border:`#facc15`},{label:`High`,min:11,max:15,fill:`#2e1200`,border:`#fb923c`},{label:`Critical`,min:16,max:1/0,fill:`#2e0a0a`,border:`#f87171`}],V=[{label:`Low`,min:1,max:5,fill:`#f0fdf4`,border:`#16a34a`},{label:`Moderate`,min:6,max:10,fill:`#fefce8`,border:`#ca8a04`},{label:`High`,min:11,max:15,fill:`#fff7ed`,border:`#ea580c`},{label:`Critical`,min:16,max:1/0,fill:`#fff1f2`,border:`#dc2626`}];function H(e,t=!0){let n=t?ne:V;return n.find(t=>e>=t.min&&e<=t.max)??n[0]}var re=e(y(),1);function U(e,t){let n=e.indexOf(`@`),r=e.indexOf(`::`);return n===-1?r===-1?{className:e,method:t??``}:{className:e.slice(0,r),method:e.slice(r+2)}:{className:e.slice(0,n),method:t??e.slice(n+1)}}function ie(e){let{className:t,method:n}=U(String(e.label??e.id),e.method),r=t.length>n.length?t:n,i=Math.max(185,Math.min(270,r.length*7.6+44));return{id:e.id,x:0,y:0,width:i,height:90,lines:[t,n].filter(Boolean),data:e}}function ae(e){if(!e.length)return;let t=0,n=0;for(let r of e)t+=r.x,n+=r.y;let r=t/e.length,i=n/e.length;for(let t of e)t.x-=r,t.y-=i}function oe(e,t,n){let r=new re.default.graphlib.Graph;r.setGraph({rankdir:n,nodesep:n===`TB`?70:50,ranksep:n===`TB`?100:120,marginx:60,marginy:60}),r.setDefaultEdgeLabel(()=>({}));for(let t of e)r.setNode(t.id,{width:t.width,height:t.height});for(let e of t)r.hasNode(e.source)&&r.hasNode(e.target)&&r.setEdge(e.source,e.target);re.default.layout(r);for(let t of e){let e=r.node(t.id);e&&(t.x=e.x,t.y=e.y)}}function se(e,t,n,r=88,i=88){let a=new Set(e.map(e=>e.id)),o=new Map,s=new Map;for(let t of e)o.set(t.id,[]),s.set(t.id,0);for(let e of t)!a.has(e.source)||!a.has(e.target)||(o.get(e.source).push(e.target),s.set(e.target,(s.get(e.target)??0)+1));let c=e.filter(e=>s.get(e.id)===0).map(e=>e.id),l=new Map,u=[...c];for(let e of c)l.set(e,0);for(;u.length;){let e=u.shift(),t=l.get(e);for(let n of o.get(e)??[]){let e=t+1;(!l.has(n)||l.get(n)<e)&&(l.set(n,e),u.push(n))}}for(let t of e)l.has(t.id)||l.set(t.id,0);let d=new Map;for(let t of e){let e=l.get(t.id);d.has(e)||d.set(e,[]),d.get(e).push(t.id)}for(let e of d.values())e.sort();for(let[,t]of d){let a=l.get(t[0]);t.forEach((o,s)=>{let c=e.find(e=>e.id===o);n===`TB`?(c.x=s*r-(t.length-1)*r/2,c.y=a*i):(c.x=a*r,c.y=s*i-(t.length-1)*i/2)})}}function W(e,t){let n=e.map(e=>Object.assign({},e)),r=new Map(n.map(e=>[e.id,e])),i=t.filter(e=>r.has(e.source)&&r.has(e.target)).map(e=>({source:e.source,target:e.target})),a=C(n).force(`link`,x(i).id(e=>e.id).distance(90)).force(`charge`,E().strength(-420)).force(`center`,D(0,0)).force(`collide`,T().radius(e=>Math.hypot(e.width,e.height)/2+14));a.stop();for(let e=0;e<450&&a.alpha()>.02;e++)a.tick();for(let t of e){let e=r.get(t.id);e&&(t.x=e.x??0,t.y=e.y??0)}}function ce(e,t){let n=e.length;n&&e.forEach((e,r)=>{let i=r/n*Math.PI*2-Math.PI/2;e.x=t*Math.cos(i),e.y=t*Math.sin(i)})}function le(e,t=200,n=120){let r=Math.ceil(Math.sqrt(Math.max(1,e.length)));e.forEach((e,i)=>{let a=Math.floor(i/r);e.x=i%r*t,e.y=a*n})}function ue(e,t,n){return e===`dagre`&&t>n?`breadthfirst`:e===`dagre`?`dagre`:e===`cose-bilkent`?`force`:e===`breadthfirst`?`breadthfirst`:e===`circle`?`circle`:e===`grid`?`grid`:`dagre`}function de(e){let t=[],n=[];for(let r of e){let e=r.data;e.source!=null&&e.target!=null?n.push({id:e.id,source:String(e.source),target:String(e.target),data:e}):t.push(ie(e))}return{nodes:t,edges:n}}var G=o();function fe(e){return Math.max(0,Math.min(255,Math.round(e))).toString(16).padStart(2,`0`)}function pe(e,t,n){return{x:n.applyX(e),y:n.applyY(t)}}function me(e,t){if(t.length===0)return{x:0,y:0};if(t.length===1||e<=0)return t[0];if(e>=1)return t[t.length-1];let n=0,r=[];for(let e=0;e<t.length-1;e++){let i=t[e+1].x-t[e].x,a=t[e+1].y-t[e].y,o=Math.sqrt(i*i+a*a);r.push(o),n+=o}if(n===0)return t[0];let i=e*n,a=0;for(let e=0;e<r.length;e++){let n=r[e];if(n!==0){if(a+n>=i||e===r.length-1){let r=(i-a)/n,o=t[e],s=t[e+1];return{x:o.x+(s.x-o.x)*r,y:o.y+(s.y-o.y)*r}}a+=n}}return t[t.length-1]}function he(e,t){let n=t.x-e.x,r=t.y-e.y,i=Math.abs(n)-(e.width+t.width)/2;return Math.abs(r)-(e.height+t.height)/2>=i?r>=0?{ex:e.x,ey:e.y+e.height/2,tx:t.x,ty:t.y-t.height/2,vertical:!0}:{ex:e.x,ey:e.y-e.height/2,tx:t.x,ty:t.y+t.height/2,vertical:!0}:n>=0?{ex:e.x+e.width/2,ey:e.y,tx:t.x-t.width/2,ty:t.y,vertical:!1}:{ex:e.x-e.width/2,ey:e.y,tx:t.x+t.width/2,ty:t.y,vertical:!1}}function ge(e,t){let{ex:n,ey:r,tx:i,ty:a,vertical:o}=he(e,t);if(o){if(Math.abs(n-i)<3||Math.abs(a-r)<=8)return[{x:n,y:r},{x:i,y:a}];let e=(r+a)/2;return[{x:n,y:r},{x:n,y:e},{x:i,y:e},{x:i,y:a}]}else{if(Math.abs(r-a)<3||Math.abs(i-n)<=8)return[{x:n,y:r},{x:i,y:a}];let e=(n+i)/2;return[{x:n,y:r},{x:e,y:r},{x:e,y:a},{x:i,y:a}]}}var _e=7;function ve(...e){return Math.max(0,Math.min(_e,...e.map(e=>e-1)))}function ye(e,t){let{ex:n,ey:r,tx:i,ty:a,vertical:o}=he(e,t);if(o){if(Math.abs(n-i)<3||Math.abs(a-r)<=8)return{d:`M${n},${r} L${i},${a}`,lx:n+6,ly:(r+a)/2,exitX:n,exitY:r,entryX:i,entryY:a};let e=(r+a)/2,t=a>r?1:-1,o=ve(Math.abs(e-r),Math.abs(a-e),Math.abs(i-n)),s=i>n?o:-o;return{d:o>0?`M${n},${r} V${e-o*t} Q${n},${e} ${n+s},${e} H${i-s} Q${i},${e} ${i},${e+o*t} V${a}`:`M${n},${r} V${e} H${i} V${a}`,lx:(n+i)/2,ly:e-14*t,exitX:n,exitY:r,entryX:i,entryY:a}}else{if(Math.abs(r-a)<3||Math.abs(i-n)<=8)return{d:`M${n},${r} L${i},${a}`,lx:(n+i)/2,ly:r-10,exitX:n,exitY:r,entryX:i,entryY:a};let e=(n+i)/2,t=i>n?1:-1,o=ve(Math.abs(e-n),Math.abs(i-e),Math.abs(a-r)),s=a>r?o:-o;return{d:o>0?`M${n},${r} H${e-o*t} Q${e},${r} ${e},${r+s} V${a-s} Q${e},${a} ${e+o*t},${a} H${i}`:`M${n},${r} H${e} V${a} H${i}`,lx:e+6*t,ly:(r+a)/2,exitX:n,exitY:r,entryX:i,entryY:a}}}function be(e,t){let n=String(e.label??``);return n?{text:n,fill:t?`rgba(255,255,255,0.4)`:`rgba(0,0,0,0.5)`,bg:t?`#111218`:`#fff`}:null}function xe(e,t,n,r,i){let a=String(e.data.type??``),o=t?L[a]??`#c9d1d9`:R[a]??`#333`,s=t?z[a]??`#0d1117`:B[a]??`#ffffff`,c=Number(e.data.metrics_cc??0)||0;if(n){let n=t?ne:V,r=n.find(e=>c>=e.min&&c<=e.max)??n[0],a=i?`#a855f7`:e.data.hasN1?`#F44336`:r.border;return{bg:r.fill,border:a,borderW:1.5,accent:r.border}}let l=t?`rgba(255,255,255,0.1)`:`rgba(0,0,0,0.12)`,u=1;return e.data.hasN1&&(l=`#F44336`,u=2),r&&(l=o,u=2),i&&(l=`#a855f7`,u=2),{bg:s,border:l,borderW:u,accent:o}}function Se({elements:e,layout:t,rankDir:n,searchQuery:r,visibleTypes:i,theme:a,onNodeSelect:o,graphRef:s,stressTestNodeId:c,stressRunKey:l,complexityOverlay:u}){let d=a===`dark`,f=d?`rgba(255,255,255,0.32)`:`rgba(0,0,0,0.38)`,p=d?`rgba(255,255,255,0.55)`:`rgba(0,0,0,0.55)`,{nodes:m,edges:h}=(0,A.useMemo)(()=>de(e),[e]),g=(0,A.useMemo)(()=>m.filter(e=>i.has(String(e.data.type))).length,[m,i]),[_,v]=(0,A.useState)(0),y=(0,A.useRef)(null),b=(0,A.useRef)(!0);(0,A.useEffect)(()=>{if(b.current){b.current=!1;return}return y.current&&window.clearTimeout(y.current),y.current=window.setTimeout(()=>{v(e=>e+1)},200),()=>{y.current&&window.clearTimeout(y.current)}},[i,t,n]);let{nodes:x,edges:C}=(0,A.useMemo)(()=>{let e=m.map(e=>({...e,lines:[...e.lines]})),r=h.map(e=>({...e})),i=ue(t,g,80);return i===`dagre`?oe(e,r,n):i===`breadthfirst`?se(e,r,n):i===`force`?W(e,r):i===`circle`?ce(e,Math.min(280,90+e.length*4)):le(e,200,130),ae(e),{nodes:e,edges:r}},[m,h,t,n,_,g]),T=(0,A.useMemo)(()=>new Map(x.map(e=>[e.id,e])),[x]),[E,D]=(0,A.useState)(new Map),k=(0,A.useRef)(null),j=(0,A.useRef)(!1),[M,N]=(0,A.useState)(new Set),[P,F]=(0,A.useState)(x);P!==x&&(F(x),D(new Map),N(new Set));let I=(0,A.useMemo)(()=>E.size===0?x:x.map(e=>{let t=E.get(e.id);return t?{...e,x:t.x,y:t.y}:e}),[x,E]),R=(0,A.useMemo)(()=>new Map(I.map(e=>[e.id,e])),[I]),z=(0,A.useRef)(R);(0,A.useEffect)(()=>{z.current=R},[R]);let B=(0,A.useCallback)(e=>i.has(String(e)),[i]),V=(0,A.useCallback)(e=>B(T.get(e.source)?.data.type)&&B(T.get(e.target)?.data.type),[T,B]),H=(0,A.useMemo)(()=>{let e=new Map;for(let t of x)e.set(t.id,[]);for(let t of C)V(t)&&e.get(t.source)?.push(t.target);let t=new Set;for(let n of M){let r=[n],i=new Set([n]);for(;r.length;){let n=r.shift();for(let a of e.get(n)??[])i.has(a)||(i.add(a),t.add(a),r.push(a))}}return t},[x,C,V,M]),re=(0,A.useMemo)(()=>{let e=new Map;for(let t of C)V(t)&&(H.has(t.target)||e.set(t.source,(e.get(t.source)??0)+1));return e},[C,V,H]),ie=(0,A.useCallback)((e,t)=>{e.stopPropagation(),N(e=>{let n=new Set(e);return n.has(t)?n.delete(t):n.add(t),n})},[]),he=(0,A.useMemo)(()=>{let e=new Map;for(let t of M){let n=0,r=new Set,i=[t];for(;i.length;){let e=i.shift();for(let t of C){if(t.source!==e||!V(t))continue;let a=t.target;r.has(a)||(r.add(a),H.has(a)&&(n++,i.push(a)))}}e.set(t,n)}return e},[M,H,C,V]),_e=(0,A.useMemo)(()=>{if(!r.trim())return null;let e=r.toLowerCase(),t=new Set;for(let n of x)String(n.data.label??n.id).toLowerCase().includes(e)&&t.add(n.id);return t},[x,r]),ve=(0,A.useMemo)(()=>{let e=new Set,t=new Set;if(!c||!T.has(c))return{nodes:e,edges:t};let n=new Set,r=new Set,i=new Set,a=[c];for(;a.length;){let e=a.shift();if(!i.has(e)){i.add(e),n.add(e);for(let t of C){if(t.source!==e||!V(t))continue;r.add(t.id);let n=t.target;i.has(n)||a.push(n)}}}return{nodes:n,edges:r}},[c,l,C,V,T]),[Se,Ce]=(0,A.useState)(new Set),[we,Te]=(0,A.useState)(null),Ee=(0,A.useCallback)(e=>{let t=new Set;for(let n of C)(n.source===e||n.target===e)&&t.add(n.id);Ce(t),Te(e),o(e)},[C,o]),De=(0,A.useCallback)(()=>{Ce(new Set),Te(null),o(null)},[o]),Oe=(0,A.useCallback)((e,t,n,r)=>{e.stopPropagation(),e.currentTarget.setPointerCapture(e.pointerId),j.current=!1,k.current={nodeId:t,startSX:e.clientX,startSY:e.clientY,origMX:n,origMY:r}},[]),ke=(0,A.useCallback)((e,t)=>{let n=k.current;if(!n||n.nodeId!==t)return;let r=e.clientX-n.startSX,i=e.clientY-n.startSY;if(!j.current&&Math.abs(r)<4&&Math.abs(i)<4)return;j.current=!0;let a=Ie.current.k;D(e=>{let o=new Map(e);return o.set(t,{x:n.origMX+r/a,y:n.origMY+i/a}),o})},[]),Ae=(0,A.useCallback)((e,t)=>{k.current?.nodeId===t&&(k.current=null)},[]),je=(0,A.useRef)(null),Me=(0,A.useRef)(null),Ne=(0,A.useRef)(null),Pe=(0,A.useRef)(null),K=(0,A.useRef)([]),q=(0,A.useRef)([]),Fe=(0,A.useRef)(0),J=(0,A.useRef)(new Map),Ie=(0,A.useRef)(w),Le=(0,A.useRef)(null),Y=(0,A.useCallback)((e,t,n=0,r=!1)=>{let i=C.find(t=>t.id===e);if(!i||!V(i))return;let a=z.current.get(i.source),o=z.current.get(i.target);if(!a||!o)return;let s=ge(a,o),c=r&&Math.random()<.65?.15+Math.random()*.55:0,l=c>0?120+Math.random()*700:0,u=r&&Math.random()<.12,d=u?.25+Math.random()*.55:0;setTimeout(()=>{K.current.push({id:`${e}-${Date.now()}-${Math.random()}`,waypoints:s,progress:0,speed:9e-4+Math.random()*4e-4,color:t,pulse:0,sparkCooldown:0,tgtNodeId:i.target,chained:r,arrived:!1,stallAt:c,stallRemaining:l,timedOut:u,timeoutAt:d})},n)},[C,V,n]),Re=(0,A.useCallback)((e,t,n=0)=>{let r=Date.now();if(r-(J.current.get(e)??0)<1800)return;J.current.set(e,r);let i=0;for(let r of C)r.source===e&&V(r)&&(Y(r.id,t,n+i*60,!0),i++)},[C,V,Y]);(0,A.useEffect)(()=>{if(!c||!T.has(c))return;let e=()=>{let e=0;for(let t of C)t.source===c&&V(t)&&(Y(t.id,`#a855f7`,e*80,!0),e++)};e();let t=window.setInterval(e,700);return()=>window.clearInterval(t)},[c,l,C,V,T,Y]),(0,A.useEffect)(()=>{let e;function t(n){e=requestAnimationFrame(t);let r=Pe.current;if(!r)return;let i=Math.min(n-Fe.current,50);Fe.current=n;let a=r.getContext(`2d`);if(!a)return;a.clearRect(0,0,r.width,r.height);let o=Ie.current,s=Math.max(.6,o.k);a.globalCompositeOperation=`lighter`;let l=[],u=x.length<=40||c,d=K.current.filter(e=>e.progress<1).length,f=Math.max(.12,1-Math.max(0,d-4)*.055);for(let e of K.current){if(!u)continue;if(e.timedOut&&e.timeoutAt>0&&e.progress>=e.timeoutAt){let t=e.waypoints.map(e=>pe(e.x,e.y,o)),n=me(e.timeoutAt,t);for(let e=0;e<18;e++){let t=e/18*Math.PI*2+Math.random()*.4,r=.06+Math.random()*.14;q.current.push({x:n.x,y:n.y,vx:Math.cos(t)*r,vy:Math.sin(t)*r,life:1,decay:.0014+Math.random()*.001,size:(1.4+Math.random()*2)*s,color:`#ef4444`})}continue}let t=e.stallAt>0&&e.progress>=e.stallAt&&e.stallRemaining>0;t?e.stallRemaining-=i:e.progress<1&&(e.progress=Math.min(1,e.progress+e.speed*f*i));let r=e.waypoints.map(e=>pe(e.x,e.y,o)),c=r[r.length-1],d=me(e.progress,r);if(!isFinite(d.x)||!isFinite(d.y)){l.push(e);continue}let p=e.stallAt>0&&e.stallRemaining>0?Math.min(1,e.stallRemaining/400):0,m=t?p>.5?`#f59e0b`:`#fb923c`:e.color;for(let t=18;t>=1;t--){let n=e.progress-t/18*.09;if(n<0)continue;let i=me(n,r),o=1-t/18,c=o*o*.55,l=(.8+o*2.6)*s;a.beginPath(),a.arc(i.x,i.y,l,0,Math.PI*2),a.fillStyle=m+fe(c*255),a.fill()}a.save(),a.shadowBlur=(t?34:24)*s,a.shadowColor=m,a.beginPath(),a.arc(d.x,d.y,5*s,0,Math.PI*2),a.fillStyle=m+`66`,a.fill(),a.restore();let h=a.createRadialGradient(d.x,d.y,0,d.x,d.y,8*s);if(h.addColorStop(0,`#ffffffee`),h.addColorStop(.35,m+`cc`),h.addColorStop(1,m+`00`),a.fillStyle=h,a.beginPath(),a.arc(d.x,d.y,8*s,0,Math.PI*2),a.fill(),t){let e=.5+.5*Math.sin(n*.012);a.beginPath(),a.arc(d.x,d.y,(10+e*8)*s,0,Math.PI*2),a.strokeStyle=`#f59e0b`+fe(e*160),a.lineWidth=1.5*s,a.stroke()}let g=1+.18*Math.sin(n*.018+e.progress*12);if(a.beginPath(),a.arc(d.x,d.y,2.2*s*g,0,Math.PI*2),a.fillStyle=`#ffffff`,a.fill(),e.progress<1&&(e.sparkCooldown-=i,e.sparkCooldown<=0)){e.sparkCooldown=35+Math.random()*40;let t=Math.random()*Math.PI*2,n=.02+Math.random()*.04;q.current.push({x:d.x,y:d.y,vx:Math.cos(t)*n,vy:Math.sin(t)*n,life:1,decay:.0028+Math.random()*.0012,size:(.8+Math.random()*1.4)*s,color:m})}if(e.progress>=1){if(!e.arrived){e.arrived=!0;for(let t=0;t<14;t++){let n=t/14*Math.PI*2+Math.random()*.3,r=.08+Math.random()*.12;q.current.push({x:c.x,y:c.y,vx:Math.cos(n)*r,vy:Math.sin(n)*r,life:1,decay:.0018+Math.random()*8e-4,size:(1.2+Math.random()*1.6)*s,color:e.color})}if(e.chained){let t=T.get(e.tgtNodeId),n=t&&L[String(t.data.type)]||e.color;Re(e.tgtNodeId,n,120)}}if(e.pulse=Math.min(1,e.pulse+.025),e.pulse<1){for(let t=0;t<3;t++){let n=e.pulse-t*.18;if(n<=0||n>=1)continue;let r=(3+n*38)*s,i=(1-n)*(1-n)*220;a.beginPath(),a.arc(c.x,c.y,r,0,Math.PI*2),a.strokeStyle=e.color+fe(i),a.lineWidth=1.5*s,a.stroke()}let t=(1-e.pulse)*(1-e.pulse)*255;a.save(),a.shadowBlur=18*s,a.shadowColor=e.color,a.beginPath(),a.arc(c.x,c.y,4*s,0,Math.PI*2),a.fillStyle=`#ffffff`+fe(t),a.fill(),a.restore(),l.push(e)}}else l.push(e)}let p=[];for(let e of q.current){if(e.x+=e.vx*i,e.y+=e.vy*i,e.vx*=.985,e.vy*=.985,e.life-=e.decay*i,e.life<=0)continue;let t=Math.max(.3,e.size*e.life);a.beginPath(),a.arc(e.x,e.y,t,0,Math.PI*2),a.fillStyle=e.color+fe(e.life*220),a.fill(),p.push(e)}q.current=p,a.globalCompositeOperation=`source-over`,K.current=l}return Fe.current=performance.now(),e=requestAnimationFrame(t),()=>cancelAnimationFrame(e)},[T,Re,x.length,c]),(0,A.useEffect)(()=>{x.length>40&&!c&&(K.current=[],q.current=[])},[x.length,c]),(0,A.useEffect)(()=>{let e=je.current,t=Pe.current;if(!e||!t)return;let n=new ResizeObserver(()=>{t.width=e.clientWidth,t.height=e.clientHeight});return n.observe(e),t.width=e.clientWidth,t.height=e.clientHeight,()=>n.disconnect()},[]),(0,A.useEffect)(()=>{let e=Me.current,t=Ne.current;if(!e||!t)return;let n=O().scaleExtent([.02,5]).filter(e=>!k.current&&(!e.ctrlKey||e.type===`wheel`)&&!e.button).on(`zoom`,e=>{Ie.current=e.transform,S(t).attr(`transform`,e.transform.toString())});return S(e).call(n),Le.current=n,()=>{S(e).on(`.zoom`,null)}},[]);let X=(0,A.useCallback)(()=>{let e=Me.current,t=je.current,n=Le.current;if(!e||!t||!n||!x.length)return;let r=1/0,i=1/0,a=-1/0,o=-1/0;for(let e of x)r=Math.min(r,e.x-e.width/2),a=Math.max(a,e.x+e.width/2),i=Math.min(i,e.y-e.height/2),o=Math.max(o,e.y+e.height/2);let s=a-r+96,c=o-i+96,l=t.clientWidth,u=t.clientHeight,d=Math.min(l/s,u/c,2)*.92,f=(r+a)/2,p=(i+o)/2,m=l/2-d*f,h=u/2-d*p,g=w.translate(m,h).scale(d);S(e).call(n.transform,g)},[x]),ze=(0,A.useCallback)(async e=>{let t=je.current;return t?(await(0,ee.default)(t,{scale:e?.scale??2,useCORS:!0,backgroundColor:d?`#0a0c10`:`#f6f7f9`})).toDataURL(`image/png`):null},[d]);(0,A.useEffect)(()=>(s.current={fit:X,toPng:ze},()=>{s.current=null}),[s,X,ze]);let Be=(0,A.useRef)(!1);return(0,A.useEffect)(()=>{Be.current=!1},[e]),(0,A.useEffect)(()=>{if(!x.length||Be.current)return;Be.current=!0;let e=requestAnimationFrame(()=>X());return()=>cancelAnimationFrame(e)},[x.length,X,e]),(0,G.jsxs)(`div`,{ref:je,style:{position:`relative`,width:`100%`,height:`100%`},children:[(0,G.jsxs)(`svg`,{ref:Me,role:`img`,"aria-label":`Execution graph`,style:{width:`100%`,height:`100%`,display:`block`,cursor:`grab`,touchAction:`none`},children:[(0,G.jsxs)(`defs`,{children:[(0,G.jsx)(`marker`,{id:`arrow-def`,markerWidth:`9`,markerHeight:`9`,refX:`8`,refY:`4.5`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,G.jsx)(`path`,{d:`M0,0 L0,9 L9,4.5 z`,fill:p})}),(0,G.jsx)(`marker`,{id:`arrow-hi`,markerWidth:`9`,markerHeight:`9`,refX:`8`,refY:`4.5`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,G.jsx)(`path`,{d:`M0,0 L0,9 L9,4.5 z`,fill:te})}),(0,G.jsx)(`marker`,{id:`arrow-st`,markerWidth:`9`,markerHeight:`9`,refX:`8`,refY:`4.5`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,G.jsx)(`path`,{d:`M0,0 L0,9 L9,4.5 z`,fill:`#a855f7`})})]}),(0,G.jsxs)(`g`,{ref:Ne,children:[(0,G.jsx)(`rect`,{x:-1e5,y:-1e5,width:2e5,height:2e5,fill:`transparent`,onClick:De,style:{pointerEvents:`all`}}),C.map(e=>{if(!V(e)||M.has(e.source)||H.has(e.source)||H.has(e.target))return null;let t=R.get(e.source),n=R.get(e.target);if(!t||!n)return null;let{d:r,lx:i,ly:a}=ye(t,n),o={x:i,y:a},s=be(e.data,d),c=Se.has(e.id),l=ve.edges.has(e.id),u=f,p=1.75,m=`url(#arrow-def)`,h=1;return l&&(u=`#a855f7`,p=2,m=`url(#arrow-st)`,h=.7),c&&(u=te,p=1.5,m=`url(#arrow-hi)`,h=1),_e&&!(_e.has(e.source)||_e.has(e.target))&&(h*=.02),(0,G.jsxs)(`g`,{children:[(0,G.jsx)(`path`,{d:r,fill:`none`,stroke:u,strokeWidth:p,strokeLinecap:`round`,strokeLinejoin:`round`,opacity:h,markerEnd:m,style:{pointerEvents:`auto`}}),s&&h>.05&&(0,G.jsx)(`g`,{transform:`translate(${o.x},${o.y})`,children:(0,G.jsx)(`text`,{textAnchor:`middle`,dominantBaseline:`middle`,fill:s.fill,fontSize:9,fontFamily:`ui-monospace, monospace`,style:{pointerEvents:`none`},children:(0,G.jsx)(`tspan`,{dx:0,dy:-8,paintOrder:`stroke fill`,stroke:s.bg,strokeWidth:6,strokeLinejoin:`round`,children:s.text})})})]},e.id)}),I.map(e=>{if(H.has(e.id))return null;let t=B(e.data.type),n=_e&&!_e.has(e.id),r=t?n?.07:1:0,i=ve.nodes.has(e.id),a=we===e.id,{bg:o,border:s,borderW:c,accent:l}=xe(e,d,u,a,i),{className:f,method:p}=U(String(e.data.label??e.id),e.data.method),m=p&&!p.includes(`(`)?p+`()`:p,h=String(e.data.type??``),g=e.width,_=g/2,v=d?`#e6edf3`:`#0d1117`,y=d?`rgba(255,255,255,0.5)`:`rgba(0,0,0,0.5)`,b=f.length>24?f.slice(0,23)+`…`:f,x=m.length>26?m.slice(0,25)+`…`:m;return(0,G.jsxs)(`g`,{transform:`translate(${e.x},${e.y})`,opacity:r,style:{pointerEvents:t&&r>.05?`auto`:`none`,cursor:`grab`},onPointerDown:t=>Oe(t,e.id,e.x,e.y),onPointerMove:t=>ke(t,e.id),onPointerUp:t=>Ae(t,e.id),onClick:t=>{t.stopPropagation(),j.current||Ee(e.id)},children:[a&&(0,G.jsx)(`rect`,{x:-_-3,y:-48,width:g+6,height:96,rx:13,fill:`none`,stroke:l,strokeWidth:6,opacity:.15}),(0,G.jsx)(`rect`,{x:-_,y:-45,width:g,height:90,rx:10,fill:o,stroke:s,strokeWidth:c,filter:e.data.hasN1&&!u?`drop-shadow(0 0 8px rgba(244,67,54,0.4))`:void 0}),(0,G.jsx)(`circle`,{cx:-_+14,cy:-27,r:4,fill:l}),(0,G.jsx)(`text`,{x:-_+24,y:-23,fontSize:10,fontFamily:`ui-monospace, monospace`,fill:l,opacity:.9,style:{pointerEvents:`none`},children:h}),!!e.data.hasN1&&(0,G.jsx)(`text`,{x:_-10,y:-23,fontSize:10,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:`#F44336`,style:{pointerEvents:`none`},children:`N+1`}),(0,G.jsx)(`text`,{x:-_+14,y:1,fontSize:13,fontWeight:700,fontFamily:`ui-sans-serif, system-ui, -apple-system, sans-serif`,fill:v,style:{pointerEvents:`none`},children:b}),x&&(0,G.jsxs)(`text`,{x:-_+14,y:19,fontSize:11,fontFamily:`ui-monospace, monospace`,fill:y,style:{pointerEvents:`none`},children:[`↻ `,x]}),(M.has(e.id)||(re.get(e.id)??0)>4)&&(0,G.jsxs)(`g`,{transform:`translate(${_+2}, 0)`,onPointerDown:e=>e.stopPropagation(),onClick:t=>ie(t,e.id),style:{cursor:`pointer`,pointerEvents:`all`},children:[(0,G.jsx)(`rect`,{x:0,y:-10,width:64,height:20,rx:10,fill:M.has(e.id)?l:d?`rgba(255,255,255,0.12)`:`rgba(0,0,0,0.10)`,stroke:l,strokeWidth:1}),(0,G.jsx)(`text`,{x:32,y:0,textAnchor:`middle`,dominantBaseline:`middle`,fill:M.has(e.id)?`#fff`:l,fontSize:10,fontWeight:700,fontFamily:`ui-monospace, monospace`,style:{pointerEvents:`none`},children:M.has(e.id)?`▶ ${he.get(e.id)??re.get(e.id)} hidden`:`▾ fold`})]})]},e.id)})]})]}),(0,G.jsx)(`canvas`,{ref:Pe,style:{position:`absolute`,top:0,left:0,pointerEvents:`none`,width:`100%`,height:`100%`}}),u&&(0,G.jsxs)(`div`,{className:`cc-legend`,children:[(0,G.jsx)(`div`,{className:`cc-legend-title`,children:`Cyclomatic Complexity`}),ne.map(e=>(0,G.jsxs)(`div`,{className:`cc-legend-row`,children:[(0,G.jsx)(`span`,{className:`cc-legend-swatch`,style:{background:e.border}}),(0,G.jsx)(`span`,{className:`cc-legend-label`,style:{color:e.border},children:e.label}),(0,G.jsx)(`span`,{className:`cc-legend-range`,children:e.max===1/0?`≥${e.min}`:`${e.min}–${e.max}`})]},e.label))]})]})}var Ce=`modulepreload`,we=function(e){return`/_laravel-brain/`+e},Te={},Ee=function(e,t,n){let r=Promise.resolve();if(t&&t.length>0){let e=document.getElementsByTagName(`link`),i=document.querySelector(`meta[property=csp-nonce]`),a=i?.nonce||i?.getAttribute(`nonce`);function o(e){return Promise.all(e.map(e=>Promise.resolve(e).then(e=>({status:`fulfilled`,value:e}),e=>({status:`rejected`,reason:e}))))}r=o(t.map(t=>{if(t=we(t,n),t in Te)return;Te[t]=!0;let r=t.endsWith(`.css`),i=r?`[rel="stylesheet"]`:``;if(n)for(let n=e.length-1;n>=0;n--){let i=e[n];if(i.href===t&&(!r||i.rel===`stylesheet`))return}else if(document.querySelector(`link[href="${t}"]${i}`))return;let o=document.createElement(`link`);if(o.rel=r?`stylesheet`:Ce,r||(o.as=`script`),o.crossOrigin=``,o.href=t,a&&o.setAttribute(`nonce`,a),document.head.appendChild(o),r)return new Promise((e,n)=>{o.addEventListener(`load`,e),o.addEventListener(`error`,()=>n(Error(`Unable to preload CSS for ${t}`)))})}))}function i(e){let t=new Event(`vite:preloadError`,{cancelable:!0});if(t.payload=e,window.dispatchEvent(t),!t.defaultPrevented)throw e}return r.then(t=>{for(let e of t||[])e.status===`rejected`&&i(e.reason);return e().catch(i)})},De=[`route`,`middleware`,`controller`,`action`,`service`,`validation_request`,`repository`,`model`,`job`,`event`,`view`,`mail`,`notification`,`enum`,`interface`,`trait`,`abstract_class`,`service_provider`,`filament_panel`,`filament_resource`,`filament_page`,`filament_page_method`,`filament_widget`,`filament_relation_manager`];function Oe(e,t){let n=[];n.push(`%%{init: {'theme': 'dark', 'themeVariables': {`),n.push(`  'background': '#0a0c10',`),n.push(`  'mainBkg': '#0d1117',`),n.push(`  'lineColor': 'rgba(255,255,255,0.35)',`),n.push(`  'edgeLabelBackground': '#111218',`),n.push(`  'edgeLabelColor': 'rgba(255,255,255,0.5)'`),n.push(`}}}%%`),n.push(`%% Laravel Brain — ${t}`),n.push(`flowchart TD`),n.push(``);let r=new Map,i=new Set,a=e=>{if(r.has(e))return r.get(e);let t=e.replace(/[^a-zA-Z0-9_]/g,`_`).replace(/^_+/,``).replace(/_+$/,``).substring(0,40);t||=`node`;let n=t,a=0;for(;i.has(n);)n=`${t}_${++a}`;return i.add(n),r.set(e,n),n},o=new Map;for(let t of e.nodes)o.has(t.type)||o.set(t.type,[]),o.get(t.type).push(t);let s=[...new Set([...De,...o.keys()])].filter(e=>(o.get(e)?.length??0)>0);for(let e of s){let t=o.get(e);n.push(`  %% ${e}`);for(let e of t){let t=a(e.id),r=ke(e);n.push(`  ${t}["${J(r)}"]`)}n.push(``)}n.push(`  %% Edges`);for(let t of e.edges){let e=a(t.source),r=a(t.target),i=t.label?`|"${J(t.label)}"| `:``;n.push(`  ${e} -->${i}${r}`)}n.push(``),n.push(`  %% Styles`);for(let e of s){let t=L[e]??`#c9d1d9`,r=z[e]??`#0d1117`;n.push(`  classDef cls_${e} fill:${r},stroke:${t},stroke-width:2px,color:#e6edf3`)}n.push(``);for(let e of s){let t=o.get(e).map(e=>a(e.id)).join(`,`);n.push(`  class ${t} cls_${e}`)}return n.join(`
-`)}function ke(e){let t=String(e.label??``),n=e.data?.method,{className:r,method:i}=U(t,n),a=i&&!i.includes(`(`)?i+`()`:i,o=[`● ${e.type}`,r];return a&&o.push(`↻ ${a}`),o.join(`
-`)}function Ae(e,t){let n=[`%% Method Flow — ${t}`,`flowchart TD`],r=0,i=()=>`s${r++}`,a=i();n.push(`  ${a}([" 🚀 ${J(t)} "])`);let o=(e,t)=>{let r=t;for(let t of e){let e=i();if(t.type===`if`){let[a,s]=[`{`,`}`];if(n.push(`  ${e}${a}"${J(t.label)}"${s}`),n.push(`  ${r} --> ${e}`),n.push(`  class ${e} cls_if`),t.then&&t.then.length>0){let r=i(),a=t.then[0];n.push(`  ${r}${K(a.type)}"${J(a.label)}"${q(a.type)}`),n.push(`  ${e} -->|"yes"| ${r}`),n.push(`  class ${r} cls_${a.type}`),o(t.then.slice(1),r)}if(t.else&&t.else.length>0){let r=i(),a=t.else[0];n.push(`  ${r}${K(a.type)}"${J(a.label)}"${q(a.type)}`),n.push(`  ${e} -->|"no"| ${r}`),n.push(`  class ${r} cls_${a.type}`),o(t.else.slice(1),r)}r=e}else if(t.type===`loop`){let i=t.n1?` ⚠️ N+1 `:``;n.push(`  ${e}[/"${i}${J(t.label)}"/]`),n.push(`  ${r} --> ${e}`),n.push(`  class ${e} ${t.n1?`cls_n1`:`cls_loop`}`),t.body&&t.body.length>0&&o(t.body,e),r=e}else{let[i,a]=[K(t.type),q(t.type)],o=Fe(t.type),s=t.n1?` ⚠️ N+1 `:``;n.push(`  ${e}${i}"${s}${o}${J(t.label)}"${a}`),n.push(`  ${r} --> ${e}`),n.push(`  class ${e} ${t.n1?`cls_n1`:`cls_${t.type}`}`),r=e}}return r};return o(e,a),n.push(``),n.push(`  %% STYLES`),n.push(`  classDef cls_call     fill:#0d47a1,stroke:#2196F3,color:#fff`),n.push(`  classDef cls_assign   fill:#212121,stroke:#616161,color:#ccc`),n.push(`  classDef cls_return   fill:#1b5e20,stroke:#4CAF50,color:#fff`),n.push(`  classDef cls_throw    fill:#b71c1c,stroke:#F44336,color:#fff`),n.push(`  classDef cls_if       fill:#f9a825,stroke:#fbc02d,color:#000`),n.push(`  classDef cls_loop     fill:#6a1b9a,stroke:#9c27b0,color:#fff`),n.push(`  classDef cls_n1       fill:#b71c1c,stroke:#ff5252,color:#fff`),n.push(`  classDef cls_dispatch fill:#bf360c,stroke:#FF5722,color:#fff`),n.push(`  classDef cls_event    fill:#0e47a1,stroke:#00BCD4,color:#fff`),n.join(`
-`)}function je(e,t){Ne(new Blob([e],{type:`text/plain`}),t)}function Me(e,t){let n=document.createElement(`a`);n.href=e,n.download=t,n.click()}function Ne(e,t){let n=URL.createObjectURL(e),r=document.createElement(`a`);r.href=n,r.download=t,r.click(),setTimeout(()=>URL.revokeObjectURL(n),2e3)}async function Pe(t,n=`#0d0f14`){let{default:r}=await Ee(async()=>{let{default:t}=await import(`./vendor-utils-DjfM5PxJ.js`).then(t=>e(t.t(),1));return{default:t}},__vite__mapDeps([0,1]));return(await r(t,{backgroundColor:n,scale:2,useCORS:!0,logging:!1})).toDataURL(`image/png`)}function K(e){switch(e){case`return`:return`([`;case`throw`:return`([`;case`dispatch`:return`[[`;case`event`:return`((`;default:return`[`}}function q(e){switch(e){case`return`:return`])`;case`throw`:return`])`;case`dispatch`:return`]]`;case`event`:return`))`;default:return`]`}}function Fe(e){switch(e){case`call`:return`→ `;case`assign`:return`= `;case`return`:return`◀ `;case`throw`:return`⚠ `;case`dispatch`:return`⚡ `;case`event`:return`📡 `;default:return``}}function J(e){return e.replace(/"/g,`'`).replace(/\n/g,`\\n`).replace(/[<>]/g,e=>e===`<`?`&lt;`:`&gt;`)}function Ie({mermaidCode:e,filename:t,title:n,onClose:r}){let[i,a]=(0,A.useState)(!1),o=(0,A.useRef)(null);return(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]),(0,G.jsx)(`div`,{className:`export-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,G.jsxs)(`div`,{className:`export-modal`,children:[(0,G.jsxs)(`div`,{className:`export-modal-header`,children:[(0,G.jsxs)(`div`,{className:`export-modal-title`,children:[(0,G.jsx)(`span`,{className:`export-modal-icon`,children:`🗺`}),(0,G.jsxs)(`div`,{children:[(0,G.jsx)(`h2`,{children:n}),(0,G.jsx)(`span`,{className:`export-modal-sub`,children:`Mermaid Flowchart`})]})]}),(0,G.jsx)(`button`,{className:`export-modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,G.jsxs)(`div`,{className:`export-modal-actions`,children:[(0,G.jsx)(`button`,{className:`export-btn export-btn--primary`,onClick:async()=>{try{await navigator.clipboard.writeText(e),a(!0),setTimeout(()=>a(!1),2e3)}catch{o.current?.select(),document.execCommand(`copy`),a(!0),setTimeout(()=>a(!1),2e3)}},children:i?`✓ Copied!`:`⎘ Copy Code`}),(0,G.jsx)(`button`,{className:`export-btn export-btn--secondary`,onClick:()=>je(e,t),children:`↓ Download .mmd`}),(0,G.jsx)(`button`,{className:`export-btn export-btn--accent`,onClick:()=>{let t=JSON.stringify({code:e,mermaid:`{}`,autoSync:!0}),n=btoa(unescape(encodeURIComponent(t)));window.open(`https://mermaid.live/edit#base64:${n}`,`_blank`)},children:`↗ Open in Mermaid Live`})]}),(0,G.jsxs)(`div`,{className:`export-modal-hint`,children:[`Paste this code at`,` `,(0,G.jsx)(`a`,{href:`https://mermaid.live`,target:`_blank`,rel:`noreferrer`,children:`mermaid.live`}),` `,`to render the diagram, or use any Mermaid-compatible tool.`]}),(0,G.jsxs)(`div`,{className:`export-code-wrapper`,children:[(0,G.jsx)(`div`,{className:`export-code-lang`,children:`mermaid`}),(0,G.jsx)(`textarea`,{ref:o,className:`export-code`,value:e,readOnly:!0,spellCheck:!1,onClick:e=>e.target.select()})]}),(0,G.jsxs)(`div`,{className:`export-modal-stats`,children:[(0,G.jsxs)(`span`,{children:[e.split(`
-`).length,` lines`]}),(0,G.jsxs)(`span`,{children:[(e.length/1024).toFixed(1),` KB`]})]})]})})}function Le({steps:e,title:t,isFatMethod:n}){let[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(!1),s=(0,A.useRef)(null);if(!e||e.length===0)return(0,G.jsx)(`div`,{className:`flowchart-empty`,children:(0,G.jsx)(`span`,{children:`No flow data available`})});let c=t??`method`;return(0,G.jsxs)(G.Fragment,{children:[n&&(0,G.jsx)(`div`,{className:`flowchart-fat-banner`,title:`Fat Method: this method exceeds complexity or line-count thresholds`,children:`🧱 Fat Method — consider breaking this into smaller methods`}),(0,G.jsxs)(`div`,{className:`flowchart-export-bar`,children:[(0,G.jsxs)(`button`,{className:`flowchart-export-btn`,onClick:async()=>{if(s.current){o(!0);try{Me(await Pe(s.current),`${c.replace(/[^a-z0-9]/gi,`_`)}_flow.png`)}finally{o(!1)}}},disabled:a,title:`Export as PNG`,children:[a?`⏳`:`🖼`,` PNG`]}),(0,G.jsx)(`button`,{className:`flowchart-export-btn`,onClick:()=>i(!0),title:`Export as Mermaid`,children:`🗺 Mermaid`})]}),(0,G.jsxs)(`div`,{className:`flowchart-root`,ref:s,children:[t&&(0,G.jsx)(`div`,{className:`flowchart-title`,children:t}),(0,G.jsx)(Y,{steps:e})]}),r&&(0,G.jsx)(Ie,{mermaidCode:Ae(e,c),filename:`${c.replace(/[^a-z0-9]/gi,`_`)}_flow.mmd`,title:c,onClose:()=>i(!1)})]})}function Y({steps:e}){return(0,G.jsx)(`div`,{className:`flowchart-list`,children:e.map((t,n)=>(0,G.jsx)(Re,{step:t,isLast:n===e.length-1},n))})}function Re({step:e,isLast:t}){return e.type===`if`?(0,G.jsxs)(`div`,{className:`flowchart-branch-wrapper`,children:[(0,G.jsx)(X,{step:e}),(0,G.jsxs)(`div`,{className:`flowchart-branches`,children:[e.then&&e.then.length>0&&(0,G.jsxs)(`div`,{className:`flowchart-branch flowchart-branch--then`,children:[(0,G.jsx)(`div`,{className:`flowchart-branch-label`,children:`then`}),(0,G.jsx)(Y,{steps:e.then})]}),e.else&&e.else.length>0&&(0,G.jsxs)(`div`,{className:`flowchart-branch flowchart-branch--else`,children:[(0,G.jsx)(`div`,{className:`flowchart-branch-label`,children:`else`}),(0,G.jsx)(Y,{steps:e.else})]})]}),!t&&(0,G.jsx)(ze,{})]}):e.type===`loop`?(0,G.jsxs)(`div`,{className:`flowchart-branch-wrapper`,children:[(0,G.jsx)(X,{step:e}),e.body&&e.body.length>0&&(0,G.jsx)(`div`,{className:`flowchart-loop-body`,children:(0,G.jsx)(Y,{steps:e.body})}),!t&&(0,G.jsx)(ze,{})]}):(0,G.jsxs)(G.Fragment,{children:[(0,G.jsx)(X,{step:e}),!t&&(0,G.jsx)(ze,{})]})}function X({step:e}){let t=`flowchart-box flowchart-box--${e.type} ${e.n1?`flowchart-box--n1`:``}`,n=Be[e.type]??``;return(0,G.jsxs)(`div`,{className:`${t} flowchart-shape--${e.type===`if`?`diamond`:e.type===`return`||e.type===`throw`?`terminal`:`rect`}`,title:e.label,children:[n&&(0,G.jsx)(`span`,{className:`flowchart-icon`,children:n}),(0,G.jsx)(`span`,{className:`flowchart-label`,children:e.label}),e.n1&&(0,G.jsx)(`span`,{className:`flowchart-n1-warn`,title:`N+1 Query Detected: This database operation is inside a loop!`,children:`⚠️ N+1`})]})}function ze(){return(0,G.jsxs)(`div`,{className:`flowchart-arrow`,children:[(0,G.jsx)(`div`,{className:`flowchart-arrow-line`}),(0,G.jsx)(`div`,{className:`flowchart-arrow-head`})]})}var Be={call:`→`,assign:`=`,return:`◀`,throw:`⚠`,if:`◆`,loop:`↻`,dispatch:`⚡`,event:`📡`};function Ve({steps:e,title:t,isFatMethod:n,onClose:r}){return(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]),(0,G.jsx)(`div`,{className:`modal-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,G.jsxs)(`div`,{className:`modal-container modal-container--large`,children:[(0,G.jsxs)(`div`,{className:`modal-header`,children:[(0,G.jsxs)(`div`,{className:`modal-title`,children:[(0,G.jsx)(`span`,{className:`modal-icon`,children:`⛓`}),(0,G.jsxs)(`div`,{children:[(0,G.jsx)(`h2`,{children:t}),(0,G.jsx)(`span`,{className:`modal-sub`,children:`Method Flow Visualization`})]})]}),(0,G.jsx)(`button`,{className:`modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,G.jsx)(`div`,{className:`modal-body flowchart-modal-body`,children:(0,G.jsx)(Le,{steps:e,isFatMethod:n})})]})})}function He(e){let[t,n]=(0,A.useState)(null),[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(null),[s,c]=(0,A.useState)(e);return e!==s&&(c(e),e?(i(!0),o(null),n(null)):(n(null),o(null))),(0,A.useEffect)(()=>{e&&fetch(`/_laravel-brain/api/source?path=${encodeURIComponent(e)}`).then(e=>e.json()).then(e=>{if(e.error)throw Error(e.error);n(e.content)}).catch(e=>o(e.message)).finally(()=>i(!1))},[e]),{content:t,loading:r,error:a}}s.registerLanguage(`php`,u);function Ue({filePath:e,highlightLine:t,theme:n}){let{content:r,loading:i,error:o}=He(e),c=(0,A.useRef)(null);(0,A.useEffect)(()=>{c.current&&c.current.scrollIntoView({block:`center`,behavior:`smooth`})},[r]);let l=e.replace(/.*\/(app|src)\//,`$1/`);return i?(0,G.jsxs)(`div`,{className:`source-state`,children:[(0,G.jsx)(`div`,{className:`loading-spinner`,style:{width:20,height:20,borderWidth:2}}),(0,G.jsx)(`span`,{children:`Loading source…`})]}):o?(0,G.jsxs)(`div`,{className:`source-state source-state--error`,children:[`Could not load file`,(0,G.jsx)(`small`,{style:{display:`block`,opacity:.6,marginTop:4},children:o})]}):r?(0,G.jsxs)(`div`,{className:`source-view`,children:[(0,G.jsx)(`div`,{className:`source-path`,title:e,children:l}),(0,G.jsx)(s,{language:`php`,style:n===`dark`?p:a,showLineNumbers:!0,wrapLines:!0,lineNumberStyle:{minWidth:`2.5em`,paddingRight:`1em`,userSelect:`none`,opacity:.4,fontSize:11},lineProps:e=>e===t?{ref:c,style:{display:`block`,backgroundColor:n===`dark`?`rgba(139,111,232,0.2)`:`rgba(139,111,232,0.12)`,borderLeft:`3px solid #8B6FE8`}}:{style:{display:`block`}},customStyle:{margin:0,padding:`12px 0`,background:`transparent`,fontSize:12,lineHeight:`1.6`,fontFamily:`ui-monospace, "Cascadia Code", monospace`},children:r})]}):null}function We({filePath:e,highlightLine:t,theme:n,onClose:r}){(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]);let i=e.split(`/`).pop()||`Source Code`;return(0,G.jsx)(`div`,{className:`modal-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,G.jsxs)(`div`,{className:`modal-container modal-container--large`,children:[(0,G.jsxs)(`div`,{className:`modal-header`,children:[(0,G.jsxs)(`div`,{className:`modal-title`,children:[(0,G.jsx)(`span`,{className:`modal-icon`,children:`📄`}),(0,G.jsxs)(`div`,{children:[(0,G.jsx)(`h2`,{children:i}),(0,G.jsx)(`span`,{className:`modal-sub`,children:e})]})]}),(0,G.jsx)(`button`,{className:`modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,G.jsx)(`div`,{className:`modal-body source-modal-body`,children:(0,G.jsx)(Ue,{filePath:e,highlightLine:t,theme:n})})]})})}var Ge=new Set([`POST`,`PUT`,`PATCH`]),Ke=new Set([`POST`,`PUT`,`PATCH`,`DELETE`]);function qe(e){let t=parseInt(e,10);return t>=200&&t<300?`#22c55e`:t>=400&&t<500?`#f97316`:t>=500?`#ef4444`:`#6b7280`}function Je(e){let t=Math.floor((Date.now()-e)/1e3);return t<60?`${t}s ago`:t<3600?`${Math.floor(t/60)}m ago`:`${Math.floor(t/3600)}h ago`}var Ye=new Map;function Z(e){let t=Ye.get(e);if(t)return t;try{let t=localStorage.getItem(`lb_st_${e}`);if(t){let n=JSON.parse(t);return Ye.set(e,n),n}}catch{}}function Xe(e,t){let n={...t,savedAt:Date.now()};Ye.set(e,n);try{localStorage.setItem(`lb_st_${e}`,JSON.stringify(n))}catch{}}function Ze(e){let t=new Set,n=[];for(let r of e.matchAll(/\{([^}?]+)(\?)?\}/g))t.has(r[1])||(n.push({name:r[1],optional:!!r[2]}),t.add(r[1]));return n}function Qe(e,t){let n=e;return n=n.replace(/\/\{([^}?]+)\?\}/g,(e,n)=>{let r=t[n]?.trim();return r?`/`+encodeURIComponent(r):``}),n=n.replace(/\{([^}?]+)\}/g,(e,n)=>encodeURIComponent(t[n]?.trim()??``)),n||`/`}function $e(e){try{let t=JSON.parse(e);return typeof t!=`object`||!t||Array.isArray(t)?null:Object.entries(t).map(([e,t])=>`${encodeURIComponent(e)}=${encodeURIComponent(String(t))}`).join(`&`)}catch{return null}}function et({method:e,uri:t,selectedId:n,onStressChange:r}){let i=`${e}::${t}`,a=Ze(t),[o,s]=(0,A.useState)(()=>{let e=window.location.href,t=e.indexOf(`/_laravel-brain`);return t===-1?window.location.origin:e.slice(0,t)}),[c,l]=(0,A.useState)(()=>Z(i)?.count??10),[u,d]=(0,A.useState)(()=>Z(i)?.concurrency??2),[f,p]=(0,A.useState)(()=>Z(i)?.headersRaw??``),[m,h]=(0,A.useState)(()=>Z(i)?.body??(Ge.has(e.toUpperCase())?`{}`:``)),[g,_]=(0,A.useState)(()=>Z(i)?.timeout??10),[v,y]=(0,A.useState)(()=>{let e=Z(i);return!!(e?.jobId&&!e?.result)}),[b,x]=(0,A.useState)(()=>Z(i)?.jobId??null),[S,C]=(0,A.useState)(()=>Z(i)?.result??null),[w,T]=(0,A.useState)(()=>Z(i)?.error??null),[E,D]=(0,A.useState)(()=>Z(i)?.routeParams??{}),[O,k]=(0,A.useState)(()=>Z(i)?.includeCsrf??Ke.has(e.toUpperCase())),[j,M]=(0,A.useState)(()=>Z(i)?.sendAsFormData??Ke.has(e.toUpperCase())),[N,P]=(0,A.useState)(0),F=(0,A.useRef)(null),I=(0,A.useRef)({result:S,error:w,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:b,routeParams:E,includeCsrf:O,sendAsFormData:j,key:i});async function ee(e,t){let n=0;for(;n<180;){if(t.aborted||(await new Promise(e=>setTimeout(e,1e3)),n++,P(n),t.aborted))return;try{let n=await(await fetch(`/_laravel-brain/api/stress-test/${e}`,{signal:t})).json();if(n.status===`done`){let e=n.result;C(e),x(null),Xe(i,{result:e,error:null,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j}),P(0);return}if(n.status===`error`){T(n.error??`Unknown error`),x(null),Xe(i,{result:null,error:n.error??`Unknown error`,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j}),P(0);return}}catch(e){if(e.name===`AbortError`)return}}P(0),x(null),Xe(i,{result:null,error:`Stress test timed out after 3 minutes`,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j}),T(`Stress test timed out after 3 minutes`)}(0,A.useEffect)(()=>{I.current={result:S,error:w,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:b,routeParams:E,includeCsrf:O,sendAsFormData:j,key:i}}),(0,A.useEffect)(()=>{let e=Z(i);e?.jobId&&!e?.result&&(r(n),F.current=new AbortController,ee(e.jobId,F.current.signal).finally(()=>{y(!1),P(0),r(null)}))},[]),(0,A.useEffect)(()=>()=>{F.current?.abort();let e=I.current;Xe(e.key,{result:e.result,error:e.error,count:e.count,concurrency:e.concurrency,headersRaw:e.headersRaw,body:e.body,timeout:e.timeout,jobId:e.jobId,routeParams:e.routeParams,includeCsrf:e.includeCsrf,sendAsFormData:e.sendAsFormData})},[]);let L=Z(i),R=L?.savedAt&&L.result?Je(L.savedAt):null;function z(e){let t={};for(let n of e.split(`
-`)){let e=n.indexOf(`:`);if(e>0){let r=n.slice(0,e).trim(),i=n.slice(e+1).trim();r&&(t[r]=i)}}return t}async function B(){let s=a.filter(e=>!e.optional&&!E[e.name]?.trim());if(s.length>0){T(`Required route param${s.length>1?`s`:``} missing: ${s.map(e=>e.name).join(`, `)}`);return}y(!0),C(null),T(null),r(n);let l=Qe(t,E),d=o.replace(/\/$/,``)+`/`+l.replace(/^\//,``);F.current=new AbortController;let p=F.current.signal,h={},_=m||null;if(Ge.has(e.toUpperCase())&&j&&m){let e=$e(m);e!==null&&(_=e,h[`Content-Type`]=`application/x-www-form-urlencoded`)}let v={...h,...z(f)};try{let t=await fetch(`/_laravel-brain/api/stress-test`,{method:`POST`,signal:p,headers:{"Content-Type":`application/json`,Accept:`application/json`},body:JSON.stringify({method:e.toUpperCase(),url:d,count:c,concurrency:u,headers:v,body:_,timeout:g,includeCsrf:Ke.has(e.toUpperCase())?O:!1})}),n=await t.json();if(!t.ok){T(n.error??`Request failed (${t.status})`);return}if(n.jobId){x(n.jobId),Xe(i,{result:null,error:null,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:n.jobId,routeParams:E,includeCsrf:O,sendAsFormData:j}),await ee(n.jobId,p);return}let r=n;C(r),x(null),Xe(i,{result:r,error:null,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j})}catch(e){e.name!==`AbortError`&&T(e instanceof Error?e.message:`Network error`)}finally{y(!1),P(0),r(null)}}let te=S?[{label:`Min`,value:`${S.timing.min}ms`},{label:`Avg`,value:`${S.timing.avg}ms`},{label:`P50`,value:`${S.timing.p50}ms`},{label:`P95`,value:`${S.timing.p95}ms`},{label:`P99`,value:`${S.timing.p99}ms`},{label:`Max`,value:`${S.timing.max}ms`},{label:`Req/s`,value:String(S.throughput)},{label:`Success`,value:`${S.successRate}%`},{label:`Wall`,value:`${S.wallTimeMs}ms`}]:[];return(0,G.jsxs)(`div`,{className:`st-section sidebar-section`,children:[(0,G.jsx)(`div`,{className:`st-toggle`,children:(0,G.jsx)(`h3`,{children:`Stress Test`})}),(0,G.jsx)(`div`,{className:`st-body`,children:(0,G.jsxs)(`div`,{className:`st-form`,children:[(0,G.jsxs)(`div`,{className:`st-form-row`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Base URL`}),(0,G.jsx)(`input`,{className:`st-input`,type:`text`,placeholder:`http://localhost:8000`,value:o,onChange:e=>s(e.target.value)})]}),(0,G.jsxs)(`div`,{className:`st-docker-hint`,children:[(0,G.jsx)(`strong`,{children:`Docker?`}),` The stress test runs `,(0,G.jsx)(`em`,{children:`inside`}),` the container — `,(0,G.jsx)(`code`,{children:`localhost:8080`}),` is the host-side port and won't be reachable there. Change Base URL to the internal service address, e.g. `,(0,G.jsx)(`code`,{children:`http://nginx`}),` or `,(0,G.jsx)(`code`,{children:`http://localhost:80`}),`.`]}),(0,G.jsxs)(`div`,{className:`st-form-row`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Target`}),(0,G.jsxs)(`span`,{className:`st-uri-preview`,children:[(0,G.jsx)(`span`,{className:`st-method-badge`,children:e.toUpperCase()}),a.length>0?Qe(t,E):t]})]}),a.length>0&&(0,G.jsxs)(`div`,{className:`st-form-col`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Route Params`}),a.map(({name:e,optional:t})=>(0,G.jsxs)(`div`,{className:`st-form-row`,style:{marginTop:4},children:[(0,G.jsxs)(`span`,{className:`st-label`,style:{minWidth:80},children:[e,t?` (opt)`:``]}),(0,G.jsx)(`input`,{className:`st-input`,type:`text`,placeholder:t?`optional`:`required`,value:E[e]??``,onChange:t=>D(n=>({...n,[e]:t.target.value}))})]},e))]}),(0,G.jsxs)(`div`,{className:`st-form-row`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Requests`}),(0,G.jsx)(`input`,{className:`st-input st-input--short`,type:`number`,min:1,max:200,value:c,onChange:e=>l(Math.max(1,Math.min(200,parseInt(e.target.value)||1)))}),(0,G.jsx)(`span`,{className:`st-label`,style:{minWidth:`auto`,marginLeft:8},children:`Concurrency`}),(0,G.jsx)(`input`,{className:`st-input st-input--short`,type:`number`,min:1,max:20,value:u,onChange:e=>d(Math.max(1,Math.min(20,parseInt(e.target.value)||1)))})]}),(0,G.jsxs)(`div`,{className:`st-form-row`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Timeout (s)`}),(0,G.jsx)(`input`,{className:`st-input st-input--short`,type:`number`,min:1,max:30,value:g,onChange:e=>_(Math.max(1,Math.min(30,parseInt(e.target.value)||10)))})]}),Ke.has(e.toUpperCase())&&(0,G.jsxs)(`div`,{className:`st-form-row`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`CSRF Token`}),(0,G.jsxs)(`label`,{style:{display:`flex`,alignItems:`center`,gap:6,cursor:`pointer`,fontSize:13},children:[(0,G.jsx)(`input`,{type:`checkbox`,checked:O,onChange:e=>k(e.target.checked)}),`Auto-inject from session`]})]}),Ge.has(e.toUpperCase())&&(0,G.jsxs)(`div`,{className:`st-form-row`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Body Format`}),(0,G.jsxs)(`label`,{style:{display:`flex`,alignItems:`center`,gap:6,cursor:`pointer`,fontSize:13},children:[(0,G.jsx)(`input`,{type:`checkbox`,checked:j,onChange:e=>M(e.target.checked)}),`Form data (application/x-www-form-urlencoded)`]})]}),(0,G.jsxs)(`div`,{className:`st-form-col`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:`Headers`}),(0,G.jsx)(`textarea`,{className:`st-textarea`,rows:3,placeholder:`Authorization: Bearer token
-Accept: application/json`,value:f,onChange:e=>p(e.target.value)})]}),Ge.has(e.toUpperCase())&&(0,G.jsxs)(`div`,{className:`st-form-col`,children:[(0,G.jsx)(`span`,{className:`st-label`,children:j?`Body (JSON → form)`:`Body (JSON)`}),(0,G.jsx)(`textarea`,{className:`st-textarea`,rows:4,placeholder:`{}`,value:m,onChange:e=>h(e.target.value)})]}),(0,G.jsx)(`button`,{className:`st-run-btn`,onClick:B,disabled:v,children:v?`⏳ Running… ${N>0?`(${N}s)`:``}`:`▶ Run Stress Test`}),R&&(0,G.jsxs)(`span`,{className:`st-last-run st-last-run--form`,children:[`Last run: `,R]}),w&&(0,G.jsx)(`div`,{className:`st-error-box`,children:w}),S&&(0,G.jsxs)(`div`,{className:`st-results`,children:[(0,G.jsx)(`div`,{className:`st-metrics-grid`,children:te.map(e=>(0,G.jsxs)(`div`,{className:`st-metric`,children:[(0,G.jsx)(`div`,{className:`st-metric-value`,children:e.value}),(0,G.jsx)(`div`,{className:`st-metric-label`,children:e.label})]},e.label))}),Object.keys(S.statusDistribution).length>0&&(0,G.jsxs)(`div`,{className:`st-dist`,children:[(0,G.jsx)(`div`,{className:`st-dist-title`,children:`Status Distribution`}),Object.entries(S.statusDistribution).sort(([e],[t])=>parseInt(e)-parseInt(t)).map(([e,t])=>(0,G.jsxs)(`div`,{className:`st-dist-row`,children:[(0,G.jsx)(`span`,{className:`st-dist-label`,children:e===`0`?`err`:e}),(0,G.jsx)(`div`,{className:`st-dist-bar-wrap`,children:(0,G.jsx)(`div`,{className:`st-dist-bar`,style:{width:`${t/S.total*100}%`,background:qe(e)}})}),(0,G.jsx)(`span`,{className:`st-dist-count`,children:t})]},e))]}),S.errors.length>0&&(0,G.jsx)(`div`,{className:`st-error-box`,style:{marginTop:8},children:S.errors.map((e,t)=>(0,G.jsx)(`div`,{children:e},t))})]})]})})]})}var tt=[`route`,`middleware`,`controller`,`action`,`validation_request`,`service`,`model`,`event`,`job`,`command`,`channel`,`schedule`,`view`,`mail`,`notification`,`enum`,`interface`,`trait`,`abstract_class`,`service_provider`];function nt(e){return e===`action`?`controller`:e}function rt(e){if(!e)return 99;let t=nt(e.type),n=tt.indexOf(t);return n===-1?99:n}function it(e){let t=e.split(`\\`),n=t[t.length-1];return n.length<=20?n:n.substring(0,18)+`…`}function at(e){let t=new Map;for(let n of e)t.has(n.source)||t.set(n.source,[]),t.get(n.source).push(n);return t}function ot(e){return e.includes(`-to-job`)||e.includes(`-to-event`)||e===`model-to-event`}function st(e,t){let n=new Map(t.nodes.map(e=>[e.id,e])),r=at(t.edges),i=new Set,a=[],o=[],s=[e];for(i.add(e);s.length>0;){let e=s.shift();a.push(e);for(let t of r.get(e)??[])o.push(t),i.has(t.target)||(i.add(t.target),s.push(t.target))}let c=[],l=new Map,u=[...a].sort((e,t)=>{let r=rt(n.get(e)),i=rt(n.get(t));return r===i?e.localeCompare(t):r-i});for(let e of u){let t=n.get(e);if(!t)continue;let r=c.length;l.set(e,r);let i=nt(t.type);c.push({id:t.id,label:it(t.label),type:i,color:L[t.type]??L[i]??`#888`})}c.unshift({id:`__client__`,label:`Client`,type:`client`,color:`#78909C`});for(let e of[...l.keys()])l.set(e,l.get(e)+1);let d=u.filter(e=>n.get(e)?.type===`model`),f=null;d.length>0&&(f=c.length,c.push({id:`__db__`,label:`Database`,type:`db`,color:`#78909C`}));let p=[],m=l.get(e);m!==void 0&&p.push({fromIndex:0,toIndex:m,label:`request`,isReturn:!1});for(let e of o){let t=l.get(e.source),n=l.get(e.target);if(t===void 0||n===void 0||t===n)continue;let r=ot(e.type);p.push({fromIndex:t,toIndex:n,label:e.label||``,isAsync:r})}if(f!==null)for(let e of d){let t=l.get(e);t!==void 0&&(p.push({fromIndex:t,toIndex:f,label:`query`,isReturn:!1}),p.push({fromIndex:f,toIndex:t,label:`result`,isReturn:!0}))}m!==void 0&&p.push({fromIndex:m,toIndex:0,label:`response`,isReturn:!0});let h=new Map,g=[];for(let e of p){let t=`${e.fromIndex}|${e.toIndex}|${e.label}|${e.isReturn?`r`:``}|${e.isAsync?`a`:``}`,n=h.get(t);if(n){n.count++;let t=e.label;g[n.idx]={...g[n.idx],label:`${t} ×${n.count}`}}else h.set(t,{idx:g.length,count:1}),g.push(e)}return{actors:c,messages:g}}function ct(e,t){let n=[`%% Sequence Diagram — ${t}`,`sequenceDiagram`,`  autonumber`];for(let t=0;t<e.actors.length;t++){let r=e.actors[t],i=`P${t}`,a=r.label.replace(/"/g,`'`);n.push(`  participant ${i} as "${a}"`)}n.push(``);for(let t of e.messages){let e=`P${t.fromIndex}`,r=`P${t.toIndex}`,i=t.label.replace(/"/g,`'`),a;a=t.isAsync?`->>`:t.isReturn?`-->>`:`->>`,n.push(`  ${e}${a}${r}: ${i}`)}return n.join(`
-`)}var lt=110,Q=52,ut=38,dt=16;function ft({diagram:e,title:t,theme:n=`dark`}){let[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(!1),s=(0,A.useRef)(null),c=n===`dark`,l=dt*2+e.actors.length*lt,u=Q+e.messages.length*ut+ut+Q,d=e=>dt+e*lt+lt/2,f=e=>Q+e*ut+ut/2,p=c?`#e0e0e0`:`#1a1a1a`,m=c?`#888`:`#999`,h=c?`rgba(255,255,255,0.10)`:`rgba(0,0,0,0.12)`,g=c?`#0d0f14`:`#ffffff`,_=c?`rgba(255,255,255,0.35)`:`rgba(0,0,0,0.30)`,v=c?`seq-arrow-dark`:`seq-arrow-light`,y=c?`seq-arrow-return-dark`:`seq-arrow-return-light`,b=c?`seq-arrow-async-dark`:`seq-arrow-async-light`,x=c?`#a0aec0`:`#555`,S=c?`#b39ddb`:`#7c4dff`;return e.actors.length===0?(0,G.jsx)(`div`,{className:`flowchart-empty`,children:(0,G.jsx)(`span`,{children:`No sequence data available`})}):(0,G.jsxs)(G.Fragment,{children:[(0,G.jsxs)(`div`,{className:`flowchart-export-bar`,children:[(0,G.jsxs)(`button`,{className:`flowchart-export-btn`,onClick:async()=>{if(s.current){o(!0);try{Me(await Pe(s.current,g),`${(t??`sequence`).replace(/[^a-z0-9]/gi,`_`)}_sequence.png`)}finally{o(!1)}}},disabled:a,title:`Export as PNG`,children:[a?`⏳`:`🖼`,` PNG`]}),(0,G.jsx)(`button`,{className:`flowchart-export-btn`,onClick:()=>i(!0),title:`Export as Mermaid`,children:`🧜 Mermaid`})]}),(0,G.jsx)(`div`,{className:`seq-diagram-root`,ref:s,children:(0,G.jsxs)(`svg`,{className:`seq-diagram-svg`,viewBox:`0 0 ${l} ${u}`,width:`100%`,style:{background:g,display:`block`},xmlns:`http://www.w3.org/2000/svg`,children:[(0,G.jsxs)(`defs`,{children:[(0,G.jsx)(`marker`,{id:v,markerWidth:`8`,markerHeight:`6`,refX:`7`,refY:`3`,orient:`auto`,children:(0,G.jsx)(`polygon`,{points:`0 0, 8 3, 0 6`,fill:x})}),(0,G.jsx)(`marker`,{id:y,markerWidth:`8`,markerHeight:`6`,refX:`7`,refY:`3`,orient:`auto`,children:(0,G.jsx)(`polyline`,{points:`0 0, 8 3, 0 6`,fill:`none`,stroke:_,strokeWidth:`1.5`})}),(0,G.jsx)(`marker`,{id:b,markerWidth:`8`,markerHeight:`6`,refX:`7`,refY:`3`,orient:`auto`,children:(0,G.jsx)(`polygon`,{points:`0 0, 8 3, 0 6`,fill:S})})]}),e.actors.map((e,t)=>{let n=d(t),r=lt-8,i=n-r/2,a=Math.floor(r/6.5),o=e.label.length>a?e.label.substring(0,a-1)+`…`:e.label;return(0,G.jsxs)(`g`,{children:[(0,G.jsx)(`rect`,{x:i,y:4,width:r,height:Q-10,rx:5,fill:c?`#1a1d24`:`#f5f5f5`,stroke:e.color,strokeWidth:1.5}),(0,G.jsx)(`text`,{x:n,y:Q/2-4,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:10.5,fontFamily:`system-ui, sans-serif`,fill:e.color,fontWeight:`600`,children:o}),(0,G.jsx)(`text`,{x:n,y:Q-12,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:8,fontFamily:`system-ui, sans-serif`,fill:m,children:e.type})]},e.id)}),e.actors.map((e,t)=>(0,G.jsx)(`line`,{x1:d(t),y1:Q,x2:d(t),y2:u-Q,stroke:h,strokeWidth:1,strokeDasharray:`4 4`},`life-${e.id}`)),e.actors.map((e,t)=>{let n=d(t),r=lt-8,i=n-r/2,a=u-Q+4,o=Math.floor(r/6.5),s=e.label.length>o?e.label.substring(0,o-1)+`…`:e.label;return(0,G.jsxs)(`g`,{children:[(0,G.jsx)(`rect`,{x:i,y:a,width:r,height:Q-10,rx:5,fill:c?`#1a1d24`:`#f5f5f5`,stroke:e.color,strokeWidth:1.5}),(0,G.jsx)(`text`,{x:n,y:a+Q/2-8,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:10.5,fontFamily:`system-ui, sans-serif`,fill:e.color,fontWeight:`600`,children:s}),(0,G.jsx)(`text`,{x:n,y:a+Q-18,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:8,fontFamily:`system-ui, sans-serif`,fill:m,children:e.type})]},`bottom-${e.id}`)}),e.messages.map((e,t)=>{let n=f(t),r=d(e.fromIndex),i=d(e.toIndex),a=i>r,o=a?r+6:r-6,s=a?i-6:i+6,c=e.isReturn===!0,l=e.isAsync===!0,u=c?_:l?S:x,h=c?`5 3`:l?`6 3`:void 0,g=c?y:l?b:v,C=(r+i)/2,w=Math.abs(i-r)-12,T=Math.max(10,Math.floor(w/6)),E=e.label.length>T?e.label.substring(0,T-1)+`…`:e.label;return(0,G.jsxs)(`g`,{children:[(0,G.jsx)(`line`,{x1:o,y1:n,x2:s,y2:n,stroke:u,strokeWidth:c?1:1.5,strokeDasharray:h,markerEnd:`url(#${g})`}),e.label&&(0,G.jsx)(`text`,{x:C,y:n-6,textAnchor:`middle`,fontSize:9,fontFamily:`system-ui, sans-serif`,fill:c?m:p,opacity:c?.75:1,children:E})]},t)})]})}),r&&(0,G.jsx)(Ie,{mermaidCode:ct(e,t??`sequence`),filename:`${(t??`sequence`).replace(/[^a-z0-9]/gi,`_`)}_sequence.mmd`,title:t??`Sequence Diagram`,onClose:()=>i(!1)})]})}function pt({diagram:e,title:t,theme:n,onClose:r}){return(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]),(0,G.jsx)(`div`,{className:`modal-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,G.jsxs)(`div`,{className:`modal-container modal-container--large`,children:[(0,G.jsxs)(`div`,{className:`modal-header`,children:[(0,G.jsxs)(`div`,{className:`modal-title`,children:[(0,G.jsx)(`span`,{className:`modal-icon`,children:`⇄`}),(0,G.jsxs)(`div`,{children:[(0,G.jsx)(`h2`,{children:t}),(0,G.jsx)(`span`,{className:`modal-sub`,children:`Sequence Diagram`})]})]}),(0,G.jsx)(`button`,{className:`modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,G.jsx)(`div`,{className:`modal-body sequence-modal-body`,children:(0,G.jsx)(ft,{diagram:e,title:t,theme:n,compact:!1})})]})})}function $({content:e,children:t,placement:a=`top`,disabled:o=!1,className:s}){let[u,p]=(0,A.useState)(!1),{refs:v,floatingStyles:y,context:x}=m({open:o?!1:u,onOpenChange:p,placement:a,middleware:[_(8),l(),i({padding:8})],whileElementsMounted:b}),{getReferenceProps:S,getFloatingProps:C}=f([n(x,{move:!1,enabled:!o,delay:{open:280,close:80}}),c(x,{enabled:!o}),d(x),r(x,{role:`tooltip`})]),w=h([v.setReference]);return(0,A.isValidElement)(t)?(0,G.jsxs)(G.Fragment,{children:[(0,A.cloneElement)(t,{ref:w,...S()}),u&&!o&&(0,G.jsx)(g,{children:(0,G.jsx)(`div`,{ref:v.setFloating,style:y,className:[`floating-tooltip`,s].filter(Boolean).join(` `),...C(),children:e})})]}):(0,G.jsx)(G.Fragment,{children:t})}var mt=200,ht=640,gt=320,_t={route:`#4CAF50`,middleware:`#FF9800`,controller:`#2196F3`,action:`#03A9F4`,service:`#9C27B0`,validation_request:`#0d9488`,model:`#F44336`,event:`#FFD600`,job:`#607D8B`,command:`#14b8a6`,channel:`#8b5cf6`,schedule:`#f97316`,view:`#ec4899`,mail:`#f472b6`,notification:`#db2777`,enum:`#0ea5e9`,interface:`#38bdf8`,trait:`#a78bfa`,abstract_class:`#94a3b8`,service_provider:`#ca8a04`,facade:`#00BCD4`,filament_panel:`#7C3AED`,filament_resource:`#A855F7`,filament_page:`#C084FC`,filament_page_method:`#E879F9`,filament_widget:`#06B6D4`,filament_relation_manager:`#0891B2`};function vt({selectedId:e,graphData:t,theme:n,onClose:r,onStressChange:i}){let[a,o]=(0,A.useState)(gt),s=(0,A.useRef)(!1),c=(0,A.useRef)(0),l=(0,A.useRef)(gt),u=(0,A.useCallback)(e=>{e.preventDefault(),s.current=!0,c.current=e.clientX,l.current=a;let t=e=>{if(!s.current)return;let t=c.current-e.clientX;o(Math.min(ht,Math.max(mt,l.current+t)))},n=()=>{s.current=!1,window.removeEventListener(`mousemove`,t),window.removeEventListener(`mouseup`,n),document.body.style.cursor=``,document.body.style.userSelect=``};document.body.style.cursor=`col-resize`,document.body.style.userSelect=`none`,window.addEventListener(`mousemove`,t),window.addEventListener(`mouseup`,n)},[a]),[d,f]=(0,A.useState)(`info`),[p,m]=(0,A.useState)(!1),[h,g]=(0,A.useState)(!1),[_,v]=(0,A.useState)(!1),[y,b]=(0,A.useState)(!1),[x,S]=(0,A.useState)(!1),[C,w]=(0,A.useState)(e);e!==C&&(w(e),f(`info`),m(!1),g(!1),v(!1),b(!1),S(!1));let T=(0,A.useMemo)(()=>{let e=new Map;return t&&t.nodes.forEach(t=>e.set(t.id,t)),e},[t]),E=(0,A.useMemo)(()=>{let e=new Map;return t&&t.edges.forEach(t=>{let n=e.get(t.target)??[];n.push(t),e.set(t.target,n)}),e},[t]),D=(0,A.useMemo)(()=>{let e=new Map;return t&&t.edges.forEach(t=>{let n=e.get(t.source)??[];n.push(t),e.set(t.source,n)}),e},[t]),O=(0,A.useMemo)(()=>!t||!e||t.nodes.find(t=>t.id===e)?.type!==`route`?null:st(e,t),[e,t]),k=(0,A.useCallback)(async()=>{if(e){S(!0);try{let t=await fetch(`/_laravel-brain/api/context?nodeId=${encodeURIComponent(e)}&budget=6000`);if(!t.ok)throw Error(`Failed to fetch context`);let n=await t.text();await navigator.clipboard.writeText(n),b(!0),setTimeout(()=>b(!1),2500)}catch{alert(`Could not copy AI context.`)}finally{S(!1)}}},[e]);if(!t)return null;if(!e)return(0,G.jsxs)(`div`,{className:`sidebar-resizable`,style:{width:a},children:[(0,G.jsx)($,{content:`Drag to resize`,children:(0,G.jsx)(`div`,{className:`sidebar-drag-handle`,onMouseDown:u})}),(0,G.jsxs)(`div`,{className:`sidebar`,children:[(0,G.jsxs)(`div`,{className:`sidebar-header`,children:[(0,G.jsx)(`h2`,{children:t.meta.project}),(0,G.jsx)(`span`,{className:`sidebar-subtitle`,children:`Laravel Lifecycle Graph`})]}),(0,G.jsxs)(`div`,{className:`sidebar-stats`,children:[(0,G.jsx)($,{content:`Total symbols in this tab's JSON graph (routes, classes, views, …).`,children:(0,G.jsxs)(`div`,{className:`stat`,children:[(0,G.jsx)(`span`,{className:`stat-value`,children:t.meta.nodeCount}),(0,G.jsx)(`span`,{className:`stat-label`,children:`Nodes`})]})}),(0,G.jsx)($,{content:`Directed links between nodes: calls, type-hints, events, views, Eloquent relations, etc.`,children:(0,G.jsxs)(`div`,{className:`stat`,children:[(0,G.jsx)(`span`,{className:`stat-value`,children:t.meta.edgeCount}),(0,G.jsx)(`span`,{className:`stat-label`,children:`Edges`})]})}),(0,G.jsx)($,{content:`HTTP route entry nodes only (subset of all node types).`,children:(0,G.jsxs)(`div`,{className:`stat`,children:[(0,G.jsx)(`span`,{className:`stat-value`,children:t.nodes.filter(e=>e.type===`route`).length}),(0,G.jsx)(`span`,{className:`stat-label`,children:`Routes`})]})})]}),(0,G.jsx)($,{content:`The inspector shows details for the selected node: metrics, flow, source, and incoming/outgoing edges.`,children:(0,G.jsx)(`p`,{className:`sidebar-hint`,children:`Click any node to inspect it`})})]})]});let j=T.get(e);if(!j)return null;let M=E.get(e)??[],N=D.get(e)??[],P=j.data?.flowSteps??[],F=j.data?.file||null,I=j.data?.line||void 0,ee=_t[j.type]??`#999`,L=j.data?.metrics,R=!!j.data?.fatMethod,z=!!j.data?.fatClass,B=!!j.data?.hasN1,te=j.data?.dbQueries??[],ne=j.data?.relationships??[],V=j.type===`middleware`&&typeof j.data?.params==`string`&&j.data.params?j.data.params.split(`,`).map(e=>e.trim()).filter(Boolean):[],H=j.data?.members??[],re=j.data?.validationRules??[],U=Object.entries(j.data??{}).filter(([e,t])=>e!==`flowSteps`&&e!==`metrics`&&e!==`fatMethod`&&e!==`fatClass`&&e!==`classMetrics`&&e!==`dbQueries`&&e!==`relationships`&&e!==`params`&&e!==`members`&&e!==`validationRules`&&!(Array.isArray(t)&&t.length===0)),ie=P.length>0||!!O,ae=!!F,oe=M.length>0||N.length>0,se=j.type===`route`,W=d===`flow`&&!ie||d===`source`&&!ae||d===`edges`&&!oe||d===`stress`&&!se?`info`:d,ce=[{id:`info`,label:`Info`,title:`Identity, type, smells, and code metrics (lines, cyclomatic complexity, …).`},...ie?[{id:`flow`,label:`Flow`,title:`Control-flow steps through this method or request (and sequence diagram for routes).`}]:[],...ae?[{id:`source`,label:`Source`,title:`Syntax-highlighted PHP source around this symbol.`}]:[],...oe?[{id:`edges`,label:`Edges`,count:M.length+N.length,title:`What calls or references this node (incoming) and what it calls (outgoing).`}]:[],...se?[{id:`stress`,label:`Stress`,title:`Send HTTP requests against this route and inspect responses (dev only).`}]:[]];return(0,G.jsxs)(`div`,{className:`sidebar-resizable`,style:{width:a},children:[(0,G.jsx)($,{content:`Drag to resize`,children:(0,G.jsx)(`div`,{className:`sidebar-drag-handle`,onMouseDown:u})}),(0,G.jsxs)(`div`,{className:`sidebar`,children:[(0,G.jsxs)(`div`,{className:`sidebar-header`,children:[(0,G.jsxs)(`div`,{className:`sidebar-header-actions`,children:[(0,G.jsx)($,{content:`Copy AI context to clipboard`,children:(0,G.jsx)(`span`,{className:`tooltip-trigger-wrap`,children:(0,G.jsx)(`button`,{type:`button`,className:`flow-popup-btn sidebar-ai-btn`,onClick:k,disabled:x,children:x?`…`:y?`✓`:`🤖`})})}),(0,G.jsx)($,{content:`Clear selection (close inspector header)`,children:(0,G.jsx)(`button`,{className:`sidebar-close`,type:`button`,onClick:r,children:`×`})})]}),(0,G.jsxs)(`div`,{className:`sidebar-badges`,children:[(0,G.jsx)($,{content:`Graph node category (route, controller, model, …). Colors match the filter panel.`,children:(0,G.jsx)(`span`,{className:`type-badge`,style:{backgroundColor:ee},children:j.type})}),typeof j.data?.visibility==`string`&&(0,G.jsxs)(`span`,{className:`visibility-badge visibility-badge--${j.data.visibility}`,children:[j.data.visibility===`public`&&`🔓 `,j.data.visibility===`protected`&&`🛡️ `,j.data.visibility===`private`&&`🔒 `,j.data.visibility]})]}),(0,G.jsx)(`h2`,{className:`sidebar-node-title`,children:j.label})]}),(R||z||B)&&(0,G.jsxs)(`div`,{className:`sidebar-smells`,children:[B&&(0,G.jsx)($,{content:`N+1 Query: database query inside a loop`,children:(0,G.jsx)(`span`,{className:`smell-badge smell-badge--n1`,children:`⚠️ N+1 Query`})}),R&&(0,G.jsx)($,{content:`Fat Method: more than 30 lines or cyclomatic complexity > 10`,children:(0,G.jsx)(`span`,{className:`smell-badge smell-badge--fat-method`,children:`🧱 Fat Method`})}),z&&(0,G.jsx)($,{content:`Fat Class: more than 10 methods or 300+ total lines`,children:(0,G.jsx)(`span`,{className:`smell-badge smell-badge--fat-class`,children:`🏗️ Fat Class`})})]}),(0,G.jsx)(`div`,{className:`sidebar-tab-bar`,children:ce.map(e=>(0,G.jsx)($,{content:e.title,children:(0,G.jsxs)(`button`,{type:`button`,className:`sidebar-tab${W===e.id?` sidebar-tab--active`:``}`,onClick:()=>f(e.id),children:[e.label,e.count!==void 0&&(0,G.jsx)(`span`,{className:`sidebar-tab-badge`,children:e.count})]})},e.id))}),(0,G.jsxs)(`div`,{className:`sidebar-tab-content`,children:[W===`info`&&(0,G.jsxs)(G.Fragment,{children:[L&&(0,G.jsxs)(`div`,{className:`sidebar-section sidebar-section--metrics`,children:[(0,G.jsx)(`h3`,{children:`Code Metrics`}),(0,G.jsxs)(`div`,{className:`metrics-grid`,children:[(0,G.jsx)($,{content:`Physical lines of code in this method (approximate, from static analysis).`,children:(0,G.jsxs)(`div`,{className:`metric-item`,children:[(0,G.jsx)(`span`,{className:`metric-value`,children:L.lineCount}),(0,G.jsx)(`span`,{className:`metric-label`,children:`Lines`})]})}),(0,G.jsx)($,{content:`Cyclomatic complexity: decision paths (branches, loops, boolean operators). Rough guide: above 10 is harder to test; above 15 is very complex.`,children:(0,G.jsxs)(`div`,{className:`metric-item`,children:[(0,G.jsx)(`span`,{className:`metric-value`,style:{color:L.cyclomaticComplexity>10?`#FF6D00`:`inherit`},children:L.cyclomaticComplexity}),(0,G.jsx)(`span`,{className:`metric-label`,children:`Complexity`})]})}),(0,G.jsx)($,{content:`Executable statements counted in this method body.`,children:(0,G.jsxs)(`div`,{className:`metric-item`,children:[(0,G.jsx)(`span`,{className:`metric-value`,children:L.statementCount}),(0,G.jsx)(`span`,{className:`metric-label`,children:`Statements`})]})}),(0,G.jsx)($,{content:`Parameters on this function or method signature.`,children:(0,G.jsxs)(`div`,{className:`metric-item`,children:[(0,G.jsx)(`span`,{className:`metric-value`,children:L.paramCount}),(0,G.jsx)(`span`,{className:`metric-label`,children:`Params`})]})})]})]}),j.type===`filament_resource`&&!!j.data?.route&&(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsx)(`h3`,{children:`Filament URL`}),(0,G.jsxs)(`div`,{className:`prop-row`,children:[(0,G.jsx)(`span`,{className:`prop-key`,children:`route`}),(0,G.jsx)(`span`,{className:`prop-value`,style:{fontFamily:`monospace`,color:`#A855F7`},children:String(j.data.route)})]})]}),ne.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsx)(`h3`,{children:`Relationships`}),ne.map((e,t)=>(0,G.jsxs)(`div`,{className:`prop-row`,children:[(0,G.jsx)(`span`,{className:`prop-key`,style:{color:`#9C27B0`},children:e.type}),(0,G.jsx)(`span`,{className:`prop-value`,children:e.related.split(`\\`).pop()??e.related})]},t))]}),V.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsx)(`h3`,{children:`ATTRIBUTES`}),V.map((e,t)=>(0,G.jsxs)(`div`,{className:`prop-row`,children:[(0,G.jsx)(`span`,{className:`prop-key`,style:{color:`#FF9800`},children:t+1}),(0,G.jsx)(`span`,{className:`prop-value`,children:e})]},t))]}),re.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section sidebar-section--validation-rules`,children:[(0,G.jsx)(`h3`,{children:`Validation rules`}),(0,G.jsx)(`ul`,{className:`sidebar-structure-list`,children:re.map((e,t)=>(0,G.jsxs)(`li`,{className:`sidebar-structure-item`,children:[(0,G.jsx)(`span`,{className:`structure-kind`,children:`field`}),(0,G.jsx)(`span`,{className:`structure-name`,children:e.field}),(0,G.jsx)(`span`,{className:`structure-value`,children:e.rules})]},t))})]}),te.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section sidebar-section--queries`,children:[(0,G.jsx)(`h3`,{children:`DB Queries`}),(0,G.jsx)(`div`,{className:`query-list`,children:te.map((e,t)=>{let n=e.table||(e.model?e.model.split(`\\`).pop():`?`);return(0,G.jsxs)(`div`,{className:`query-item`,children:[(0,G.jsx)(`span`,{className:`query-op query-op--${[`insert`,`update`,`delete`,`statement`].includes(e.operation)?`write`:`read`}`,children:e.operation}),(0,G.jsx)(`span`,{className:`query-table`,title:e.model||void 0,children:n}),e.type===`raw`&&(0,G.jsx)(`span`,{className:`query-badge query-badge--raw`,children:`SQL`})]},t)})})]}),H.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsx)(`h3`,{children:`Structure`}),(0,G.jsx)(`ul`,{className:`sidebar-structure-list`,children:H.map((e,t)=>(0,G.jsxs)(`li`,{className:`sidebar-structure-item`,children:[(0,G.jsx)(`span`,{className:`structure-kind`,children:String(e.kind??`item`)}),(0,G.jsx)(`span`,{className:`structure-name`,children:String(e.name??``)}),typeof e.declaringClass==`string`&&e.declaringClass!==``&&(0,G.jsx)(`span`,{className:`structure-decl`,title:`Declared on parent class`,children:e.declaringClass}),e.value!==void 0&&e.value!==null&&(0,G.jsx)(`span`,{className:`structure-value`,children:String(e.value)}),e.static===!0&&(0,G.jsx)(`span`,{className:`structure-flag`,children:`static`}),typeof e.visibility==`string`&&(0,G.jsx)(`span`,{className:`structure-vis`,children:e.visibility})]},t))})]}),(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsx)(`h3`,{children:`Properties`}),U.map(([e,t])=>(0,G.jsxs)(`div`,{className:`prop-row`,children:[(0,G.jsx)(`span`,{className:`prop-key`,children:e}),(0,G.jsx)(`span`,{className:`prop-value`,children:Array.isArray(t)?t.map(e=>typeof e==`object`&&e?Object.values(e).join(` `):String(e)).join(`, `)||`—`:String(t)||`—`})]},e))]})]}),W===`flow`&&(0,G.jsxs)(G.Fragment,{children:[P.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section sidebar-section--flowchart`,children:[(0,G.jsxs)(`div`,{className:`sidebar-section-header`,children:[(0,G.jsx)(`h3`,{children:`Method Flow`}),(0,G.jsx)(`button`,{className:`flow-popup-btn`,title:`Open in large view`,onClick:()=>m(!0),children:`⤢`})]}),(0,G.jsx)(Le,{steps:P,isFatMethod:R}),p&&(0,G.jsx)(Ve,{steps:P,title:j.label,isFatMethod:R,onClose:()=>m(!1)})]}),O&&(0,G.jsxs)(`div`,{className:`sidebar-section sidebar-section--sequence`,children:[(0,G.jsxs)(`div`,{className:`sidebar-section-header`,children:[(0,G.jsx)(`h3`,{children:`Sequence Diagram`}),(0,G.jsx)(`button`,{className:`flow-popup-btn`,title:`Open in large view`,onClick:()=>v(!0),children:`⤢`})]}),(0,G.jsx)(ft,{diagram:O,title:j.label,theme:n}),_&&(0,G.jsx)(pt,{diagram:O,title:j.label,theme:n,onClose:()=>v(!1)})]})]}),W===`source`&&F&&(0,G.jsxs)(`div`,{className:`sidebar-section sidebar-section--source`,children:[(0,G.jsxs)(`div`,{className:`sidebar-section-header`,children:[(0,G.jsx)(`h3`,{children:`Source Code`}),(0,G.jsx)(`button`,{className:`flow-popup-btn`,title:`Open in large view`,onClick:()=>g(!0),children:`⤢`})]}),(0,G.jsx)(Ue,{filePath:F,highlightLine:I,theme:n}),h&&(0,G.jsx)(We,{filePath:F,highlightLine:I,theme:n,onClose:()=>g(!1)})]}),W===`edges`&&(0,G.jsxs)(G.Fragment,{children:[N.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsxs)(`h3`,{children:[`Outgoing (`,N.length,`)`]}),N.map(e=>{let t=T.get(e.target);return(0,G.jsxs)(`div`,{className:`edge-row`,children:[(0,G.jsx)(`span`,{className:`edge-label`,children:e.label}),(0,G.jsx)(`span`,{className:`edge-target`,children:t?.label??e.target})]},e.id)})]}),M.length>0&&(0,G.jsxs)(`div`,{className:`sidebar-section`,children:[(0,G.jsxs)(`h3`,{children:[`Incoming (`,M.length,`)`]}),M.map(e=>(0,G.jsxs)(`div`,{className:`edge-row`,children:[(0,G.jsx)(`span`,{className:`edge-target`,children:T.get(e.source)?.label??e.source}),(0,G.jsx)(`span`,{className:`edge-label`,children:e.label})]},e.id))]})]}),W===`stress`&&se&&e&&(0,G.jsx)(et,{method:String(j.data?.method??`GET`),uri:String(j.data?.uri??`/`),theme:n,selectedId:e,onStressChange:i},e)]})]})]})}var yt=[{id:`claude`,label:`Claude Code`,path:`CLAUDE.md`,icon:`🟠`,description:`Anthropic Claude Code CLI & IDE`},{id:`cursor`,label:`Cursor`,path:`.cursor/rules/laravel-brain.mdc`,icon:`⬛`,description:`Cursor AI editor (MDC format with frontmatter)`},{id:`windsurf`,label:`Windsurf`,path:`.windsurf/rules/laravel-brain.md`,icon:`🌊`,description:`Windsurf by Codeium`},{id:`copilot`,label:`GitHub Copilot`,path:`.github/copilot-instructions.md`,icon:`🐙`,description:`Applied repo-wide automatically`},{id:`junie`,label:`JetBrains Junie`,path:`.junie/guidelines.md`,icon:`🧠`,description:`JetBrains AI assistant`},{id:`aider`,label:`Aider`,path:`CONVENTIONS.md`,icon:`⌨️`,description:`Load with: aider --read CONVENTIONS.md`},{id:`agents`,label:`AGENTS.md`,path:`AGENTS.md`,icon:`🌐`,description:`Universal open standard — 60+ tools`},{id:`codex`,label:`OpenAI Codex`,path:`CODEX.md`,icon:`🟢`,description:`Load with: codex --context CODEX.md`}];function bt({onClose:e}){let[t,n]=(0,A.useState)(new Set(yt.map(e=>e.id))),[r,i]=(0,A.useState)({}),[a,o]=(0,A.useState)(!1),[s,c]=(0,A.useState)(null),l=(0,A.useCallback)(e=>{n(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})},[]),u=(0,A.useCallback)(()=>n(new Set(yt.map(e=>e.id))),[]),d=(0,A.useCallback)(()=>n(new Set),[]),f=(0,A.useCallback)(async e=>{o(!0),c(null);let n={};t.forEach(e=>{n[e]={status:`generating`}}),i(n);try{let n=await fetch(`/_laravel-brain/api/generate-rules`,{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({targets:[...t],force:e})}),r=await n.json();if(n.status===409&&r.existing){c(r.existing);let e={};t.forEach(t=>{e[t]={status:`idle`}}),i(e);return}if(!n.ok){let e=r.error??`Generation failed`,n={};t.forEach(t=>{n[t]={status:`error`,error:e}}),i(n);return}let a={};for(let e of r.results??[])a[e.target]=e.success?{status:`success`,path:e.path}:{status:`error`,error:e.error??`Unknown error`};i(a)}catch{let e={};t.forEach(t=>{e[t]={status:`error`,error:`Network error`}}),i(e)}finally{o(!1)}},[t]),p=(0,A.useCallback)(()=>f(!1),[f]),m=(0,A.useCallback)(()=>f(!0),[f]),h=(0,A.useCallback)(()=>c(null),[]),g=Object.values(r).filter(e=>e.status===`success`).length,_=Object.values(r).filter(e=>e.status===`error`).length,v=g+_>0;return(0,G.jsx)(`div`,{className:`export-overlay`,onClick:t=>{t.target===t.currentTarget&&e()},children:(0,G.jsxs)(`div`,{className:`export-modal ai-rules-modal`,children:[(0,G.jsxs)(`div`,{className:`export-modal-header`,children:[(0,G.jsxs)(`div`,{className:`export-modal-title`,children:[(0,G.jsx)(`span`,{className:`export-modal-icon`,children:`🤖`}),(0,G.jsxs)(`div`,{children:[(0,G.jsx)(`h2`,{children:`Generate AI Rules Files`}),(0,G.jsx)(`div`,{className:`export-modal-sub`,children:`Write context files for AI coding assistants into your project`})]})]}),(0,G.jsx)(`button`,{className:`export-modal-close`,onClick:e,children:`×`})]}),s&&(0,G.jsxs)(`div`,{className:`ai-rules-overwrite-banner`,children:[(0,G.jsx)(`div`,{className:`ai-rules-overwrite-icon`,children:`⚠️`}),(0,G.jsxs)(`div`,{className:`ai-rules-overwrite-body`,children:[(0,G.jsxs)(`strong`,{children:[`The following file`,s.length===1?``:`s`,` already exist`,s.length===1?`s`:``,`:`]}),(0,G.jsx)(`ul`,{className:`ai-rules-overwrite-list`,children:s.map(e=>(0,G.jsx)(`li`,{children:(0,G.jsx)(`code`,{children:e.path})},e.target))}),(0,G.jsxs)(`span`,{children:[`Do you want to overwrite `,s.length===1?`it`:`them`,`?`]})]}),(0,G.jsxs)(`div`,{className:`ai-rules-overwrite-actions`,children:[(0,G.jsx)(`button`,{className:`export-btn export-btn--secondary`,onClick:h,children:`Cancel`}),(0,G.jsx)(`button`,{className:`export-btn export-btn--danger`,onClick:m,children:`Overwrite`})]})]}),(0,G.jsxs)(`div`,{className:`ai-rules-select-bar`,children:[(0,G.jsxs)(`span`,{className:`ai-rules-select-label`,children:[t.size,` of `,yt.length,` selected`]}),(0,G.jsx)(`button`,{className:`ai-rules-select-link`,onClick:u,children:`All`}),(0,G.jsx)(`span`,{className:`ai-rules-select-sep`,children:`·`}),(0,G.jsx)(`button`,{className:`ai-rules-select-link`,onClick:d,children:`None`})]}),(0,G.jsx)(`div`,{className:`ai-rules-grid`,children:yt.map(e=>{let n=t.has(e.id),i=r[e.id];return(0,G.jsxs)(`label`,{className:`ai-rules-card ${n?`ai-rules-card--selected`:``} ${a?`ai-rules-card--disabled`:``}`,children:[(0,G.jsx)(`input`,{type:`checkbox`,className:`ai-rules-checkbox`,checked:n,disabled:a,onChange:()=>l(e.id)}),(0,G.jsx)(`span`,{className:`ai-rules-card-icon`,children:e.icon}),(0,G.jsxs)(`div`,{className:`ai-rules-card-body`,children:[(0,G.jsx)(`span`,{className:`ai-rules-card-label`,children:e.label}),(0,G.jsx)(`code`,{className:`ai-rules-card-path`,children:e.path}),(0,G.jsx)(`span`,{className:`ai-rules-card-desc`,children:e.description})]}),(0,G.jsxs)(`div`,{className:`ai-rules-card-status`,children:[i?.status===`generating`&&(0,G.jsx)(`span`,{className:`ai-rules-status ai-rules-status--spinning`,children:`⏳`}),i?.status===`success`&&(0,G.jsx)(`span`,{className:`ai-rules-status ai-rules-status--ok`,title:i.path,children:`✓`}),i?.status===`error`&&(0,G.jsx)(`span`,{className:`ai-rules-status ai-rules-status--err`,title:i.error,children:`✗`})]})]},e.id)})}),v&&(0,G.jsxs)(`div`,{className:`ai-rules-summary`,children:[g>0&&(0,G.jsxs)(`span`,{className:`ai-rules-summary--ok`,children:[`✓ `,g,` file`,g===1?``:`s`,` written`]}),_>0&&(0,G.jsxs)(`span`,{className:`ai-rules-summary--err`,children:[`✗ `,_,` error`,_===1?``:`s`]})]}),(0,G.jsxs)(`div`,{className:`ai-rules-footer`,children:[(0,G.jsx)(`button`,{className:`export-btn export-btn--secondary`,onClick:e,disabled:a,children:v?`Close`:`Cancel`}),(0,G.jsx)(`button`,{className:`export-btn export-btn--primary ${a?`export-btn--loading`:``}`,onClick:p,disabled:a||t.size===0,children:a?(0,G.jsxs)(G.Fragment,{children:[(0,G.jsx)(`span`,{className:`btn-spinner btn-spinner--small`}),` Generating…`]}):`Generate ${t.size>0?t.size:``} File${t.size===1?``:`s`}`})]})]})})}function xt(e){let t=Math.floor(e/1e3);if(t<60)return`${t}s`;let n=Math.floor(t/60);if(n<60)return`${n}m`;let r=Math.floor(n/60);return r<24?`${r}h`:`${Math.floor(r/24)}d`}var St=[{value:`dagre`,label:`Hierarchical (dagre)`},{value:`breadthfirst`,label:`Breadth-first`},{value:`cose-bilkent`,label:`Force (cose-bilkent)`},{value:`circle`,label:`Circle`},{value:`grid`,label:`Grid`}];function Ct({label:e,icon:t,children:n,buttonTitle:r}){let[i,a]=(0,A.useState)(!1),o=(0,A.useRef)(null);return(0,A.useEffect)(()=>{let e=e=>{o.current&&!o.current.contains(e.target)&&a(!1)};return document.addEventListener(`mousedown`,e,!0),()=>document.removeEventListener(`mousedown`,e,!0)},[]),(0,G.jsxs)(`div`,{className:`action-dropdown`,ref:o,children:[r?(0,G.jsx)($,{content:r,children:(0,G.jsxs)(`button`,{type:`button`,className:`toolbar-btn ${i?`toolbar-btn--active`:``}`,onClick:()=>a(!i),children:[t,` `,(0,G.jsx)(`span`,{children:e}),` `,(0,G.jsx)(`span`,{className:`dropdown-chevron`,children:`▾`})]})}):(0,G.jsxs)(`button`,{type:`button`,className:`toolbar-btn ${i?`toolbar-btn--active`:``}`,onClick:()=>a(!i),children:[t,` `,(0,G.jsx)(`span`,{children:e}),` `,(0,G.jsx)(`span`,{className:`dropdown-chevron`,children:`▾`})]}),i&&(0,G.jsx)(`div`,{className:`action-dropdown-menu`,children:n})]})}function wt({layout:e,rankDir:t,onRankDirChange:n,nodeCount:r,edgeCount:i,visibleCount:a,activeTabLabel:o,graphData:s,analyzedAt:c,theme:l,onLayoutChange:u,onSearch:d,onToggleTheme:f,graphRef:p,complexityOverlay:m,onToggleComplexityOverlay:h}){let[g,_]=(0,A.useState)(``),[v,y]=(0,A.useState)(!1),[b,x]=(0,A.useState)(!1),[S,C]=(0,A.useState)(!1),w=(0,A.useRef)(null);(0,A.useEffect)(()=>(w.current&&clearTimeout(w.current),w.current=setTimeout(()=>{d(g)},250),()=>{w.current&&clearTimeout(w.current)}),[g,d]);let T=()=>p.current?.fit(),E=()=>{p.current?.toPng({scale:2}).then(e=>{e&&Me(e,`${o.replace(/[^a-z0-9]/gi,`_`)}_graph.png`)})},D=()=>{s&&y(!0)},O=async()=>{if(window.confirm(`This will re-scan the entire project. Proceed?`)){C(!0);try{(await fetch(`/_laravel-brain/api/scan`,{method:`POST`})).ok?window.location.reload():alert(`Scan failed.`)}catch{alert(`Scan failed.`)}finally{C(!1)}}},[k,j]=(0,A.useState)(()=>Date.now());(0,A.useEffect)(()=>{let e=setInterval(()=>j(Date.now()),6e4);return()=>clearInterval(e)},[]);let M=(0,A.useMemo)(()=>{if(!c)return null;let e=k-new Date(c).getTime();return{ageMs:e,isStale:e>24*3600*1e3,label:`Scanned ${xt(e)} ago`}},[c,k]);return(0,G.jsxs)(G.Fragment,{children:[(0,G.jsxs)(`div`,{className:`toolbar`,children:[(0,G.jsx)($,{content:`Laravel Brain — static analysis graph of routes, classes, and dependencies.`,children:(0,G.jsx)(`div`,{className:`toolbar-brand`,children:(0,G.jsx)(`img`,{src:`/_laravel-brain/logo.png`,alt:`Laravel Brain`,className:`toolbar-logo-img`,width:38,height:38,decoding:`async`})})}),(0,G.jsxs)(`div`,{className:`toolbar-stats`,children:[(0,G.jsx)($,{content:`Nodes are classes, routes, views, and other symbols in this graph. Visible count respects sidebar type filters.`,children:(0,G.jsxs)(`span`,{className:`stat-chip`,children:[a,`/`,r,` nodes`]})}),(0,G.jsx)($,{content:`Edges are references between nodes (calls, type-hints, dispatch, views, etc.).`,children:(0,G.jsxs)(`span`,{className:`stat-chip`,children:[i,` edges`]})}),r>80&&(0,G.jsx)($,{content:`Large graph: dagre auto-switched to breadthfirst`,children:(0,G.jsx)(`span`,{className:`stat-chip stat-chip--warn`,children:`⚠ large graph`})}),M&&(M.isStale?(0,G.jsx)($,{content:`Graph is over 24h old (last scanned ${c?new Date(c).toLocaleString():``}). Click to re-scan.`,children:(0,G.jsxs)(`button`,{type:`button`,className:`stat-chip stat-chip--stale`,onClick:O,children:[`⚠ `,M.label,` — Re-scan?`]})}):(0,G.jsx)($,{content:`Last scanned: ${c?new Date(c).toLocaleString():``}`,children:(0,G.jsx)(`span`,{className:`stat-chip stat-chip--age`,children:M.label})}))]}),(0,G.jsxs)(`div`,{className:`toolbar-controls`,children:[(0,G.jsx)(`div`,{className:`toolbar-group`,children:(0,G.jsxs)(Ct,{label:`View`,icon:(0,G.jsx)(`span`,{children:`⊡`}),buttonTitle:`Arrange the graph: layout algorithm, orientation, and zoom to fit.`,children:[(0,G.jsxs)(`div`,{className:`dropdown-item`,children:[(0,G.jsx)(`label`,{children:`Layout`}),(0,G.jsx)($,{content:`How node positions are computed. Try breadth-first or force layouts if hierarchical (dagre) is cluttered.`,children:(0,G.jsx)(`select`,{value:e,onChange:e=>u(e.target.value),className:`toolbar-select`,children:St.map(({value:e,label:t})=>(0,G.jsx)(`option`,{value:e,children:t},e))})})]}),e===`dagre`&&(0,G.jsxs)(`div`,{className:`dropdown-item`,children:[(0,G.jsx)(`label`,{children:`Orientation`}),(0,G.jsx)($,{content:`For hierarchical layout: draw graph left-to-right or top-to-bottom.`,children:(0,G.jsx)(`button`,{type:`button`,className:`toolbar-btn toolbar-btn--rank`,onClick:()=>n(t===`LR`?`TB`:`LR`),children:t===`LR`?`Horizontal (LR)`:`Vertical (TB)`})})]}),(0,G.jsx)(`div`,{className:`dropdown-item`,children:(0,G.jsx)($,{content:`Zoom and pan so all visible nodes fit in the viewport.`,children:(0,G.jsxs)(`button`,{type:`button`,onClick:T,className:`toolbar-btn w-full`,children:[(0,G.jsx)(`span`,{children:`⊡`}),` `,(0,G.jsx)(`span`,{children:`Fit to Screen`})]})})})]})}),(0,G.jsx)(`div`,{className:`toolbar-group`,children:(0,G.jsx)($,{content:`Cyclomatic complexity: number of independent paths through code (branches, loops). Higher values often mean harder-to-test methods. Colors nodes by this metric when enabled.`,children:(0,G.jsxs)(`button`,{type:`button`,className:`toolbar-btn ${m?`toolbar-btn--active`:``}`,onClick:h,children:[(0,G.jsx)(`span`,{children:`◈`}),` `,(0,G.jsx)(`span`,{children:`Complexity`})]})})}),(0,G.jsx)(`div`,{className:`toolbar-group`,children:(0,G.jsx)(`div`,{className:`toolbar-search-wrapper`,children:(0,G.jsx)($,{content:`Filter highlighted matches in the graph by class name, path, or label.`,children:(0,G.jsx)(`input`,{type:`search`,placeholder:`Search nodes...`,className:`toolbar-search`,value:g,onChange:e=>_(e.target.value)})})})}),(0,G.jsx)(`div`,{className:`toolbar-group`,children:(0,G.jsxs)(Ct,{label:`Export`,icon:(0,G.jsx)(`span`,{children:`🖼`}),buttonTitle:`Save the graph as an image, Mermaid diagram text, or AI context rules.`,children:[(0,G.jsx)(`div`,{className:`dropdown-item`,children:(0,G.jsx)($,{content:`Raster image of the current graph viewport (useful for docs or slides).`,children:(0,G.jsxs)(`button`,{type:`button`,onClick:E,className:`toolbar-btn w-full`,children:[(0,G.jsx)(`span`,{children:`🖼`}),` `,(0,G.jsx)(`span`,{children:`Download PNG`})]})})}),(0,G.jsx)(`div`,{className:`dropdown-item`,children:(0,G.jsx)($,{content:`Mermaid flowchart syntax you can paste into Markdown, Notion, or mermaid.live.`,children:(0,G.jsx)(`span`,{className:`tooltip-trigger-wrap tooltip-trigger-wrap--block`,children:(0,G.jsxs)(`button`,{type:`button`,onClick:D,className:`toolbar-btn w-full`,disabled:!s,children:[(0,G.jsx)(`span`,{children:`🗺`}),` `,(0,G.jsx)(`span`,{children:`Copy Mermaid Code`})]})})})}),(0,G.jsx)(`div`,{className:`dropdown-item`,children:(0,G.jsx)($,{content:`Build a rules snippet describing this project for AI coding assistants.`,children:(0,G.jsxs)(`button`,{type:`button`,onClick:()=>x(!0),className:`toolbar-btn w-full`,children:[(0,G.jsx)(`span`,{children:`🤖`}),` `,(0,G.jsx)(`span`,{children:`Generate AI Rules`})]})})})]})}),(0,G.jsxs)(`div`,{className:`toolbar-group`,children:[(0,G.jsx)($,{content:`Re-run project scan to refresh routes, graphs, and analysis from disk.`,children:(0,G.jsx)(`span`,{className:`tooltip-trigger-wrap tooltip-trigger-wrap--block`,children:(0,G.jsx)(`button`,{type:`button`,onClick:O,className:`toolbar-btn toolbar-btn--scan ${S?`toolbar-btn--loading`:``}`,disabled:S,"aria-busy":S,children:S?(0,G.jsxs)(G.Fragment,{children:[(0,G.jsx)(`div`,{className:`btn-spinner btn-spinner--small`,"aria-hidden":!0}),(0,G.jsx)(`span`,{children:`Scanning…`})]}):(0,G.jsxs)(G.Fragment,{children:[(0,G.jsx)(`span`,{className:`toolbar-scan__glyph`,"aria-hidden":!0,children:(0,G.jsxs)(`svg`,{width:`16`,height:`16`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,G.jsx)(`path`,{d:`M21 12a9 9 0 0 0-9-9 9.75 9.75 0 0 0-6.74 2.74L3 8`}),(0,G.jsx)(`path`,{d:`M3 3v5h5`}),(0,G.jsx)(`path`,{d:`M3 12a9 9 0 0 0 9 9 9.75 9.75 0 0 0 6.74-2.74L21 16`}),(0,G.jsx)(`path`,{d:`M16 16h5v5`})]})}),(0,G.jsx)(`span`,{className:`toolbar-scan__label`,children:`Re-scan`})]})})})}),(0,G.jsx)($,{content:l===`dark`?`Switch to light mode`:`Switch to dark mode`,children:(0,G.jsx)(`button`,{type:`button`,onClick:f,className:`theme-toggle`,children:l===`dark`?`☀️`:`🌙`})})]})]})]}),b&&(0,G.jsx)(bt,{onClose:()=>x(!1)}),v&&s&&(0,G.jsx)(Ie,{mermaidCode:Oe(s,o),filename:`${o.replace(/[^a-z0-9]/gi,`_`)}_graph.mmd`,title:`${o} — Full Lifecycle Graph`,onClose:()=>y(!1)})]})}var Tt=[{type:`route`,label:`Routes`,color:`#4CAF50`,tip:`HTTP route entry points (method + URI) into your application.`},{type:`middleware`,label:`Middleware`,color:`#FF9800`,tip:`Pipeline classes that wrap requests before they reach controllers.`},{type:`controller`,label:`Controllers`,color:`#2196F3`,tip:`Controller classes that orchestrate HTTP requests.`},{type:`livewire_component`,label:`Livewire Components`,color:`#FB7185`,tip:`Livewire components that extend Livewire\\Component.`},{type:`action`,label:`Actions`,color:`#03A9F4`,tip:`Controller methods or invokable actions that run for a route.`},{type:`service`,label:`Services`,color:`#9C27B0`,tip:`Injected services and domain/helper classes resolved from the container.`},{type:`validation_request`,label:`Validation requests`,color:`#0d9488`,tip:`Form request classes that validate and authorize input.`},{type:`model`,label:`Models`,color:`#F44336`,tip:`Eloquent models and data-layer classes.`},{type:`event`,label:`Events`,color:`#FFD600`,tip:`Domain events dispatched on the event bus.`},{type:`job`,label:`Jobs`,color:`#607D8B`,tip:`Queueable jobs for background work.`},{type:`command`,label:`Commands`,color:`#14b8a6`,tip:`Artisan console commands.`},{type:`channel`,label:`Channels`,color:`#8b5cf6`,tip:`Broadcast channels for realtime messaging.`},{type:`schedule`,label:`Schedules`,color:`#f97316`,tip:`Scheduled task definitions (cron / scheduler).`},{type:`view`,label:`Views`,color:`#ec4899`,tip:`Blade templates and view names returned from controllers.`},{type:`mail`,label:`Mail`,color:`#f472b6`,tip:`Mailable classes.`},{type:`notification`,label:`Notifications`,color:`#db2777`,tip:`Notification classes (mail, database, etc.).`},{type:`enum`,label:`Enums`,color:`#0ea5e9`,tip:`PHP backed enums referenced from analyzed code.`},{type:`interface`,label:`Interfaces`,color:`#38bdf8`,tip:`PHP interfaces implemented or type-hinted in this graph.`},{type:`trait`,label:`Traits`,color:`#a78bfa`,tip:`PHP traits used by classes in this lifecycle.`},{type:`abstract_class`,label:`Abstract classes`,color:`#94a3b8`,tip:`Abstract base classes extended in this path.`},{type:`service_provider`,label:`Service providers`,color:`#ca8a04`,tip:`Service providers that register bindings and boot logic.`},{type:`filament_panel`,label:`Filament Panels`,color:`#7C3AED`,tip:`Filament panel configuration (admin area).`},{type:`filament_resource`,label:`Filament Resources`,color:`#A855F7`,tip:`Filament CRUD resources for Eloquent models.`},{type:`filament_page`,label:`Filament Pages`,color:`#C084FC`,tip:`Custom Filament pages (non-resource screens).`},{type:`filament_page_method`,label:`Page Methods`,color:`#E879F9`,tip:`Methods on Filament pages (actions, hooks).`},{type:`filament_widget`,label:`Filament Widgets`,color:`#06B6D4`,tip:`Filament dashboard or resource widgets.`},{type:`filament_relation_manager`,label:`Relation Managers`,color:`#0891B2`,tip:`Filament relation managers on resource edit pages.`}];function Et({visibleTypes:e,counts:t,onToggle:n,onShowAll:r,onHideAll:i}){return(0,G.jsxs)(`div`,{className:`filter-panel`,children:[(0,G.jsxs)(`div`,{className:`filter-header`,children:[(0,G.jsx)($,{content:`Show or hide node categories in the graph. Edges to hidden types may disappear.`,children:(0,G.jsx)(`span`,{className:`filter-title`,children:`Filters`})}),(0,G.jsxs)(`div`,{className:`filter-actions`,children:[(0,G.jsx)($,{content:`Show every node type that appears in this graph.`,children:(0,G.jsx)(`button`,{type:`button`,onClick:r,className:`filter-link`,children:`all`})}),(0,G.jsx)(`span`,{className:`filter-sep`,children:`·`}),(0,G.jsx)($,{content:`Hide all types (graph may be empty until you re-enable some).`,children:(0,G.jsx)(`button`,{type:`button`,onClick:i,className:`filter-link`,children:`none`})})]})]}),Tt.map(({type:r,label:i,color:a,tip:o})=>{let s=t[r]??0;if(s===0)return null;let c=e.has(r);return(0,G.jsx)($,{content:o,children:(0,G.jsxs)(`label`,{className:`filter-item ${c?``:`filter-item--dim`}`,children:[(0,G.jsx)(`input`,{type:`checkbox`,checked:c,onChange:()=>n(r),className:`filter-checkbox`}),(0,G.jsx)(`span`,{className:`filter-dot`,style:{backgroundColor:a}}),(0,G.jsx)(`span`,{className:`filter-label`,children:i}),(0,G.jsx)(`span`,{className:`filter-count`,children:s})]})},r)})]})}var Dt=new Set([`action`,`service`,`validation_request`,`controller`,`job`,`command`]),Ot={all:`All analyzed methods with cyclomatic complexity above 0 (actions, services, controllers, jobs, commands).`,complex:`Methods with cyclomatic complexity greater than 10 (increasing risk of bugs and hard tests).`,critical:`Methods with cyclomatic complexity greater than 15 (very high branching — strong refactor candidate).`};function kt({graphData:e,filter:t,onFilterChange:n,onNodeSelect:r,selectedId:i}){let a=(0,A.useMemo)(()=>e?e.nodes.filter(e=>Dt.has(e.type)).map(e=>({node:e,cc:e.data?.metrics?.cyclomaticComplexity??0})).filter(e=>t===`complex`?e.cc>10:t===`critical`?e.cc>15:e.cc>0).sort((e,t)=>t.cc-e.cc):[],[e,t]);return(0,G.jsxs)(`div`,{className:`complexity-panel`,children:[(0,G.jsx)(`div`,{className:`complexity-filters`,children:[`all`,`complex`,`critical`].map(e=>(0,G.jsx)($,{content:Ot[e],children:(0,G.jsx)(`button`,{type:`button`,className:`complexity-filter-btn ${t===e?`complexity-filter-btn--active`:``}`,onClick:()=>n(e),children:e===`all`?`All`:e===`complex`?`>10`:`>15`})},e))}),(0,G.jsx)($,{content:`Cyclomatic complexity counts decision points (if, loop, case, &&, ||, catch, etc.) in each method. Sorted highest first.`,children:(0,G.jsxs)(`div`,{className:`complexity-summary`,children:[a.length,` method`,a.length===1?``:`s`]})}),!e&&(0,G.jsx)(`div`,{className:`complexity-empty`,children:`Select a route tab to see complexity data.`}),e&&a.length===0&&(0,G.jsx)(`div`,{className:`complexity-empty`,children:`No methods match the current filter.`}),(0,G.jsx)(`div`,{className:`complexity-list`,children:a.map(e=>{let t=H(e.cc),n=L[e.node.type]??`#94a3b8`;return(0,G.jsx)($,{content:`Cyclomatic complexity ${e.cc}. Click to focus this node in the graph.`,children:(0,G.jsxs)(`button`,{type:`button`,className:`complexity-row ${i===e.node.id?`complexity-row--active`:``}`,onClick:()=>r(e.node.id),children:[(0,G.jsx)(`span`,{className:`complexity-badge`,style:{color:t.border,borderColor:t.border},children:e.cc}),(0,G.jsx)(`span`,{className:`complexity-label`,children:e.node.label}),(0,G.jsx)(`span`,{className:`complexity-type`,style:{color:n},children:e.node.type})]})},e.node.id)})})]})}var At=160,jt=480,Mt=240,Nt={GET:`#4ade80`,POST:`#60a5fa`,PUT:`#f59e0b`,PATCH:`#a78bfa`,DELETE:`#f87171`};function Pt({tab:e,isActive:t,isLoading:n,onSelect:r}){let[i,...a]=e.label.split(` `),o=i in Nt,s=o?i:null,c=o?a.join(` `):e.label,l=s?Nt[s]:`#94a3b8`;return(0,G.jsx)($,{content:`Open this lifecycle graph. Nodes: ${e.nodeCount} (symbols in the graph). Edges: ${e.edgeCount} (references between them).`,children:(0,G.jsxs)(`button`,{className:`left-nav-item ${t?`left-nav-item--active`:``}`,type:`button`,onClick:()=>r(e),children:[s?(0,G.jsx)(`span`,{className:`left-nav-method`,style:{color:l},children:s}):(0,G.jsx)(`span`,{className:`left-nav-method`,style:{color:l},children:`›`}),(0,G.jsx)(`span`,{className:`left-nav-uri`,children:c}),n&&(0,G.jsx)(`span`,{className:`left-nav-badge`,children:`…`})]})})}function Ft(e){return e.tabs.length+e.subGroups.reduce((e,t)=>e+Ft(t),0)}function It(e,t){let n=e.tabs.filter(e=>e.label.toLowerCase().includes(t)),r=e.subGroups.map(e=>It(e,t)).filter(e=>e!==null);return n.length===0&&r.length===0?null:{...e,tabs:n,subGroups:r}}function Lt({group:e,depth:t,prefixKey:n,activeId:r,loadingId:i,onSelect:a,expandedPrefixes:o,togglePrefix:s}){let c=o.has(n),l=Ft(e),u=t===0?14:12;return(0,G.jsxs)(`div`,{className:`left-nav-prefix-group`,style:t>0?{marginLeft:`${t*12}px`}:void 0,children:[(0,G.jsx)($,{content:`URL segment /${e.prefix} — routes that share this path prefix.`,children:(0,G.jsxs)(`button`,{type:`button`,className:`left-nav-prefix-header`,style:t>0?{fontSize:`10.5px`,opacity:.85}:void 0,onClick:()=>s(n),children:[(0,G.jsx)(`span`,{className:`left-nav-prefix-chevron`,children:c?`▾`:`▸`}),(0,G.jsx)(`span`,{className:`left-nav-prefix-icon`,children:(0,G.jsx)(`svg`,{width:u,height:u,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:(0,G.jsx)(`path`,{d:`M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z`})})}),(0,G.jsxs)(`span`,{className:`left-nav-prefix-name`,children:[`/`,e.prefix]}),(0,G.jsx)(`span`,{className:`left-nav-prefix-count`,children:l})]})}),c&&(0,G.jsxs)(G.Fragment,{children:[e.tabs.map(e=>(0,G.jsx)(Pt,{tab:e,isActive:e.id===r,isLoading:e.id===i,onSelect:a},e.id)),e.subGroups.map(e=>(0,G.jsx)(Lt,{group:e,depth:t+1,prefixKey:`${n}::${e.prefix}`,activeId:r,loadingId:i,onSelect:a,expandedPrefixes:o,togglePrefix:s},e.prefix))]})]})}function Rt({fileGroups:e,activeId:t,loadingId:n,onSelect:r,visibleTypes:i,counts:a,onToggle:o,onShowAll:s,onHideAll:c,graphData:l,complexityFilter:u,onComplexityFilterChange:d,onNodeSelect:f,selectedId:p}){let[m,h]=(0,A.useState)(`routes`),[g,_]=(0,A.useState)(320),[v,y]=(0,A.useState)(Mt),[b,x]=(0,A.useState)(``),S=(0,A.useRef)(null),C=(0,A.useRef)(0),w=(0,A.useRef)(0),T=(0,A.useRef)(!1),E=(0,A.useRef)(0),D=(0,A.useRef)(Mt),O=(0,A.useCallback)(e=>{e.preventDefault(),T.current=!0,E.current=e.clientX,D.current=v;let t=e=>{if(!T.current)return;let t=e.clientX-E.current;y(Math.min(jt,Math.max(At,D.current+t)))},n=()=>{T.current=!1,window.removeEventListener(`mousemove`,t),window.removeEventListener(`mouseup`,n),document.body.style.cursor=``,document.body.style.userSelect=``};document.body.style.cursor=`col-resize`,document.body.style.userSelect=`none`,window.addEventListener(`mousemove`,t),window.addEventListener(`mouseup`,n)},[v]),[k,j]=(0,A.useState)(new Set),[M,N]=(0,A.useState)(new Set),P=e=>j(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n}),F=e=>N(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n}),I=(0,A.useCallback)(e=>{e.preventDefault(),C.current=e.clientY,w.current=g;let t=e=>{let t=S.current;if(!t)return;let n=e.clientY-C.current,r=t.clientHeight;_(Math.max(80,Math.min(r-80,w.current+n)))},n=()=>{window.removeEventListener(`mousemove`,t),window.removeEventListener(`mouseup`,n)};window.addEventListener(`mousemove`,t),window.addEventListener(`mouseup`,n)},[g]),ee=(0,A.useMemo)(()=>{let t=b.trim().toLowerCase();return t?e.map(e=>({...e,flatTabs:e.flatTabs.filter(e=>e.label.toLowerCase().includes(t)),prefixGroups:e.prefixGroups.map(e=>It(e,t)).filter(e=>e!==null)})).filter(e=>e.prefixGroups.length>0||e.flatTabs.length>0):e},[e,b]);return(0,G.jsxs)(`div`,{className:`left-sidebar-resizable`,style:{width:v},children:[(0,G.jsxs)(`div`,{className:`left-sidebar`,ref:S,children:[(0,G.jsxs)(`div`,{className:`left-sidebar-tabs`,children:[(0,G.jsx)($,{content:`Browse routes and commands grouped by file. Selecting one loads its lifecycle graph.`,children:(0,G.jsx)(`button`,{type:`button`,className:`left-sidebar-tab ${m===`routes`?`left-sidebar-tab--active`:``}`,onClick:()=>h(`routes`),children:`Routes`})}),(0,G.jsx)($,{content:`List methods ranked by cyclomatic complexity for the current graph.`,children:(0,G.jsx)(`button`,{type:`button`,className:`left-sidebar-tab ${m===`complexity`?`left-sidebar-tab--active`:``}`,onClick:()=>h(`complexity`),children:`Complexity`})})]}),(0,G.jsx)(`div`,{className:`left-sidebar-top`,style:{height:g},children:m===`complexity`?(0,G.jsx)(kt,{graphData:l,filter:u,onFilterChange:d,onNodeSelect:f,selectedId:p}):(0,G.jsxs)(G.Fragment,{children:[(0,G.jsxs)(`div`,{className:`left-nav-search`,children:[(0,G.jsx)($,{content:`Filter the route list by HTTP method, URI segment, or command name (substring match).`,children:(0,G.jsx)(`input`,{className:`left-nav-search-input`,type:`text`,placeholder:`Search routes…`,value:b,onChange:e=>x(e.target.value)})}),b&&(0,G.jsx)($,{content:`Clear route search`,children:(0,G.jsx)(`button`,{type:`button`,className:`left-nav-search-clear`,onClick:()=>x(``),children:`×`})})]}),(0,G.jsx)(`div`,{className:`left-nav`,children:ee.map(e=>{let i=k.has(e.fileName),a=e.flatTabs.length+e.prefixGroups.reduce((e,t)=>e+Ft(t),0);return(0,G.jsxs)(`div`,{className:`left-nav-file-group`,children:[(0,G.jsx)($,{content:`Route definitions from ${e.fileName}. Expand to see URI groups and endpoints.`,children:(0,G.jsxs)(`button`,{type:`button`,className:`left-nav-file-header`,onClick:()=>P(e.fileName),children:[(0,G.jsx)(`span`,{className:`left-nav-file-chevron`,children:i?`▾`:`▸`}),(0,G.jsx)(`span`,{className:`left-nav-file-icon`,children:(0,G.jsxs)(`svg`,{width:`14`,height:`14`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,G.jsx)(`path`,{d:`M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z`}),(0,G.jsx)(`polyline`,{points:`14 2 14 8 20 8`})]})}),(0,G.jsx)(`span`,{className:`left-nav-file-name`,children:e.fileName}),(0,G.jsx)(`span`,{className:`left-nav-file-count`,children:a})]})}),i&&(0,G.jsxs)(G.Fragment,{children:[e.flatTabs.map(e=>(0,G.jsx)(Pt,{tab:e,isActive:e.id===t,isLoading:e.id===n,onSelect:r},e.id)),e.prefixGroups.map(i=>(0,G.jsx)(Lt,{group:i,depth:0,prefixKey:`${e.fileName}::${i.prefix}`,activeId:t,loadingId:n,onSelect:r,expandedPrefixes:M,togglePrefix:F},i.prefix))]})]},e.fileName)})})]})}),(0,G.jsx)(`div`,{className:`left-sidebar-handle`,onMouseDown:I}),(0,G.jsx)(`div`,{className:`left-sidebar-bottom`,children:(0,G.jsx)(Et,{visibleTypes:i,counts:a,onToggle:o,onShowAll:s,onHideAll:c})})]}),(0,G.jsx)($,{content:`Drag to resize`,children:(0,G.jsx)(`div`,{className:`left-sidebar-drag-handle`,onMouseDown:O})})]})}var zt=`route.middleware.controller.livewire_component.action.service.validation_request.model.event.job.command.channel.schedule.view.mail.notification.enum.interface.trait.abstract_class.service_provider.filament_panel.filament_resource.filament_page.filament_page_method.filament_widget.filament_relation_manager`.split(`.`);function Bt(){let{theme:e,toggle:t}=I(),{manifest:n,loading:r,error:i}=M(),{state:a,elements:o,load:s}=P(),c=F(o),[l,u]=(0,A.useState)(null),[d,f]=(0,A.useState)(null),[p,m]=(0,A.useState)(`dagre`),[h,g]=(0,A.useState)(null),[_,v]=(0,A.useState)(``),[y,b]=(0,A.useState)(new Set(zt)),[x,S]=(0,A.useState)(`TB`),[C,w]=(0,A.useState)(null),[T,E]=(0,A.useState)(0),D=(0,A.useRef)(null),O=(0,A.useCallback)(e=>{if(l?.id===e.id)return;let t=new URL(window.location.href);t.searchParams.get(`tab`)!==e.id&&(t.searchParams.set(`tab`,e.id),window.history.pushState({tabId:e.id},``,t.toString())),u(e),f(e.id),v(``),s(e.file)},[l,s]),[k,j]=(0,A.useState)(n);if(n!==k&&(j(n),n&&!l)){let e=new URLSearchParams(window.location.search).get(`tab`),t=n.tabs.find(t=>t.id===e);t&&O(t)}let[N,ee]=(0,A.useState)(a.data);if(a.data!==N){if(ee(a.data),a.data){let e=a.data.nodes.length;b(new Set(e>80?zt.filter(e=>e!==`middleware`):zt))}g(null)}(0,A.useEffect)(()=>{let e=()=>{if(!n)return;let e=new URLSearchParams(window.location.search).get(`tab`),t=n.tabs.find(t=>t.id===e);t&&(u(t),s(t.file))};return window.addEventListener(`popstate`,e),()=>window.removeEventListener(`popstate`,e)},[n,s]);let L=(0,A.useCallback)(e=>{g(e)},[]),[R,z]=(0,A.useState)(a.loading);a.loading!==R&&(z(a.loading),a.loading||f(null));let B=(0,A.useMemo)(()=>{if(!n)return{fileGroups:[]};let e=n.tabs,t=new Map;e.forEach(e=>{let n;if(e.category===`Command`)n=`Console Commands`;else if(e.category===`Channel`)n=`Broadcast Channels`;else if(e.category===`Schedule`)n=`Schedules`;else if(e.category===`Filament`){let t=e.panelId??``;n=`Filament · ${t?t.charAt(0).toUpperCase()+t.slice(1)+` Panel`:`Filament`}`}else n=e.routeFile??`routes.php`;let r=t.get(n)??[];r.push(e),t.set(n,r)});let r=new Set([`GET`,`POST`,`PUT`,`PATCH`,`DELETE`,`OPTIONS`,`ANY`]);function i(e){return[...e.entries()].sort(([e],[t])=>e.localeCompare(t)).map(([e,t])=>({prefix:e,tabs:t.tabs,subGroups:i(t.children)}))}return{fileGroups:[...t.entries()].sort(([e],[t])=>e.localeCompare(t)).map(([e,t])=>{let n=[],a=new Map;return t.forEach(e=>{let t=e.label.split(` `)[0];if(!r.has(t)){n.push(e);return}let i=(e.uriPrefix??e.label.replace(/^[A-Z]+\s+/,``).split(`/`).filter(Boolean)[0]??``).split(`/`).filter(Boolean);if(i.length===0){n.push(e);return}let o=a;for(let t=0;t<i.length;t++){let n=i[t];o.has(n)||o.set(n,{tabs:[],children:new Map}),t===i.length-1?o.get(n).tabs.push(e):o=o.get(n).children}}),{fileName:e,prefixGroups:i(a),flatTabs:n}})}},[n]),te=(0,A.useMemo)(()=>a.data?a.data.nodes.reduce((e,t)=>(e[t.type]=(e[t.type]??0)+1,e),{}):{},[a.data]),ne=(0,A.useMemo)(()=>a.data?a.data.nodes.filter(e=>y.has(e.type)).length:0,[a.data,y]),V=(0,A.useCallback)(e=>{b(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})},[]),H=(0,A.useCallback)(()=>b(new Set(zt)),[]),re=(0,A.useCallback)(()=>b(new Set),[]),[U,ie]=(0,A.useState)(!1),[ae,oe]=(0,A.useState)(!1),[se,W]=(0,A.useState)(`all`);return r?(0,G.jsxs)(`div`,{className:`loading-screen`,children:[(0,G.jsx)(`div`,{className:`loading-spinner`}),(0,G.jsx)(`p`,{children:`Loading project graph...`})]}):i||!n?(0,G.jsx)(`div`,{className:`error-screen welcome-screen`,children:(0,G.jsxs)(`div`,{className:`welcome-card`,children:[(0,G.jsx)(`div`,{className:`welcome-icon`,children:(0,G.jsx)(`img`,{src:`/_laravel-brain/logo.png`,alt:`Laravel Brain`})}),(0,G.jsx)(`h2`,{children:`Welcome to Laravel Brain`}),(0,G.jsx)(`p`,{children:`No project analysis found. To begin exploring your code architecture, please run an initial scan.`}),i&&i!==`HTTP 404`&&(0,G.jsx)(`div`,{className:`error-details`,children:(0,G.jsxs)(`small`,{children:[`Error: `,i]})}),(0,G.jsx)(`button`,{className:`scan-btn ${U?`scan-btn--loading`:``}`,onClick:async()=>{if(window.confirm(`This will scan the entire project. Proceed?`)){ie(!0);try{(await fetch(`/_laravel-brain/api/scan`,{method:`POST`})).ok?window.location.reload():alert(`Scan failed.`)}catch{alert(`Scan failed.`)}finally{ie(!1)}}},disabled:U,children:U?(0,G.jsxs)(G.Fragment,{children:[(0,G.jsx)(`div`,{className:`btn-spinner`}),`Analyzing Project...`]}):`🚀 Start Initial Scan`}),(0,G.jsxs)(`div`,{className:`welcome-hint`,children:[`Alternatively, run `,(0,G.jsx)(`code`,{children:`php artisan brain:scan`}),` in your terminal.`]})]})}):(0,G.jsxs)(`div`,{className:`app`,children:[(0,G.jsx)(wt,{layout:p,nodeCount:a.data?.meta.nodeCount??l?.nodeCount??0,edgeCount:a.data?.meta.edgeCount??l?.edgeCount??0,visibleCount:ne,activeTabLabel:l?.label??`graph`,graphData:a.data??null,analyzedAt:n.analyzedAt,theme:e,onLayoutChange:m,rankDir:x,onRankDirChange:S,onSearch:v,onToggleTheme:t,graphRef:D,complexityOverlay:ae,onToggleComplexityOverlay:()=>oe(e=>!e)}),(0,G.jsxs)(`div`,{className:`main`,children:[(0,G.jsx)(Rt,{fileGroups:B.fileGroups??[],activeId:l?.id??null,loadingId:d,onSelect:O,visibleTypes:y,counts:te,onToggle:V,onShowAll:H,onHideAll:re,graphData:a.data??null,complexityFilter:se,onComplexityFilterChange:W,onNodeSelect:L,selectedId:h}),(0,G.jsxs)(`div`,{className:`graph-container`,children:[a.loading&&(0,G.jsxs)(`div`,{className:`graph-loading-overlay`,children:[(0,G.jsx)(`div`,{className:`loading-spinner`}),(0,G.jsxs)(`p`,{children:[`Loading `,l?.label,`…`]})]}),a.error&&(0,G.jsx)(`div`,{className:`graph-loading-overlay`,children:(0,G.jsxs)(`p`,{style:{color:`#F44336`},children:[`Error: `,a.error]})}),!l&&!a.loading&&(0,G.jsx)($,{content:`Pick a route or command in the left sidebar to load its dependency graph.`,children:(0,G.jsxs)(`div`,{className:`graph-placeholder`,children:[(0,G.jsx)(`div`,{className:`placeholder-icon`,children:(0,G.jsx)(`svg`,{width:`64`,height:`64`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`1.5`,strokeLinecap:`round`,strokeLinejoin:`round`,children:(0,G.jsx)(`polyline`,{points:`22 12 18 12 15 21 9 3 6 12 2 12`})})}),(0,G.jsx)(`h3`,{children:`Select a route to explore`}),(0,G.jsx)(`p`,{children:`Expand the files in the sidebar and choose a route or command to visualize its execution lifecycle and dependencies.`})]})}),!a.loading&&l&&c.length===0&&!a.error&&(0,G.jsx)($,{content:`This endpoint produced no analyzable nodes. It may be a closure, a redirect-only route, or outside the scanner’s rules.`,children:(0,G.jsxs)(`div`,{className:`graph-placeholder`,children:[(0,G.jsx)(`div`,{className:`placeholder-icon`,children:(0,G.jsxs)(`svg`,{width:`64`,height:`64`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`1.5`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,G.jsx)(`circle`,{cx:`12`,cy:`12`,r:`10`}),(0,G.jsx)(`line`,{x1:`12`,y1:`8`,x2:`12`,y2:`12`}),(0,G.jsx)(`line`,{x1:`12`,y1:`16`,x2:`12.01`,y2:`16`})]})}),(0,G.jsx)(`h3`,{children:`Empty Graph`}),(0,G.jsx)(`p`,{children:`No nodes or edges found for this route.`})]})}),!a.loading&&c.length>0&&(0,G.jsx)(Se,{elements:c,layout:p,searchQuery:_,rankDir:x,visibleTypes:y,theme:e,onNodeSelect:L,graphRef:D,stressTestNodeId:C,stressRunKey:T,complexityOverlay:ae},l?.id)]}),(0,G.jsx)(vt,{selectedId:h,graphData:a.data,theme:e,onClose:()=>g(null),onStressChange:e=>{w(e),e!==null&&E(e=>e+1)}})]})]})}(0,j.createRoot)(document.getElementById(`root`)).render((0,G.jsx)(A.StrictMode,{children:(0,G.jsx)(Bt,{})}));
